Minfied the actual files for better load times Added src directory for development
