+++ src/data.js
@@ -1,0 +1,6750 @@
+const data = {
+  Dhaka: [
+    {
+      city: "Demra",
+      sub: "Demra",
+      postalCode: "1360",
+    },
+    {
+      city: "Demra",
+      sub: "Sarulia",
+      postalCode: "1361",
+    },
+    {
+      city: "Demra",
+      sub: "Matuail",
+      postalCode: "1362",
+    },
+    {
+      city: "Dhaka Cantt.",
+      sub: "Dhaka CantonmentTSO",
+      postalCode: "1206",
+    },
+    {
+      city: "Dhamrai",
+      sub: "Dhamrai",
+      postalCode: "1350",
+    },
+    {
+      city: "Dhamrai",
+      sub: "Kamalpur",
+      postalCode: "1351",
+    },
+    {
+      city: "Dhanmondi",
+      sub: "Jigatala TSO",
+      postalCode: "1209",
+    },
+    {
+      city: "Gulshan",
+      sub: "Gulshan Model Town",
+      postalCode: "1212",
+    },
+    {
+      city: "Gulshan",
+      sub: "Banani TSO",
+      postalCode: "1213",
+    },
+    {
+      city: "Jatrabari",
+      sub: "Dhania TSO",
+      postalCode: "1232",
+    },
+    {
+      city: "Joypara",
+      sub: "Joypara",
+      postalCode: "1330",
+    },
+    {
+      city: "Joypara",
+      sub: "Palamganj",
+      postalCode: "1331",
+    },
+    {
+      city: "Joypara",
+      sub: "Narisha",
+      postalCode: "1332",
+    },
+    {
+      city: "Keraniganj",
+      sub: "Keraniganj",
+      postalCode: "1310",
+    },
+    {
+      city: "Keraniganj",
+      sub: "Dhaka Jute Mills",
+      postalCode: "1311",
+    },
+    {
+      city: "Keraniganj",
+      sub: "Ati",
+      postalCode: "1312",
+    },
+    {
+      city: "Keraniganj",
+      sub: "Kalatia",
+      postalCode: "1313",
+    },
+    {
+      city: "Khilgaon",
+      sub: "KhilgaonTSO",
+      postalCode: "1219",
+    },
+    {
+      city: "Khilkhet",
+      sub: "KhilkhetTSO",
+      postalCode: "1229",
+    },
+    {
+      city: "Lalbag",
+      sub: "Posta TSO",
+      postalCode: "1211",
+    },
+    {
+      city: "Mirpur",
+      sub: "Mirpur TSO",
+      postalCode: "1218",
+    },
+    {
+      city: "Mohammadpur",
+      sub: "Mohammadpur Housing",
+      postalCode: "1207",
+    },
+    {
+      city: "Mohammadpur",
+      sub: "Sangsad BhabanTSO",
+      postalCode: "1225",
+    },
+    {
+      city: "Motijheel",
+      sub: "BangabhabanTSO",
+      postalCode: "1222",
+    },
+    {
+      city: "Motijheel",
+      sub: "DilkushaTSO",
+      postalCode: "1223",
+    },
+    {
+      city: "Nawabganj",
+      sub: "Nawabganj",
+      postalCode: "1320",
+    },
+    {
+      city: "Nawabganj",
+      sub: "Hasnabad",
+      postalCode: "1321",
+    },
+    {
+      city: "Nawabganj",
+      sub: "Daudpur",
+      postalCode: "1322",
+    },
+    {
+      city: "Nawabganj",
+      sub: "Agla",
+      postalCode: "1323",
+    },
+    {
+      city: "Nawabganj",
+      sub: "Khalpar",
+      postalCode: "1324",
+    },
+    {
+      city: "Nawabganj",
+      sub: "Churain",
+      postalCode: "1325",
+    },
+    {
+      city: "New market",
+      sub: "New Market TSO",
+      postalCode: "1205",
+    },
+    {
+      city: "Palton",
+      sub: "Dhaka GPO",
+      postalCode: "1000",
+    },
+    {
+      city: "Ramna",
+      sub: "Shantinagr TSO",
+      postalCode: "1217",
+    },
+    {
+      city: "Sabujbag",
+      sub: "Basabo TSO",
+      postalCode: "1214",
+    },
+    {
+      city: "Savar",
+      sub: "Savar",
+      postalCode: "1340",
+    },
+    {
+      city: "Savar",
+      sub: "Dairy Farm",
+      postalCode: "1341",
+    },
+    {
+      city: "Savar",
+      sub: "Jahangirnagar Univer",
+      postalCode: "1342",
+    },
+    {
+      city: "Savar",
+      sub: "Saver P.A.T.C",
+      postalCode: "1343",
+    },
+    {
+      city: "Savar",
+      sub: "Savar Canttonment",
+      postalCode: "1344",
+    },
+    {
+      city: "Savar",
+      sub: "Shimulia",
+      postalCode: "1345",
+    },
+    {
+      city: "Savar",
+      sub: "Kashem Cotton Mills",
+      postalCode: "1346",
+    },
+    {
+      city: "Savar",
+      sub: "Rajphulbaria",
+      postalCode: "1347",
+    },
+    {
+      city: "Savar",
+      sub: "Amin Bazar",
+      postalCode: "1348",
+    },
+    {
+      city: "Savar",
+      sub: "EPZ",
+      postalCode: "1349",
+    },
+    {
+      city: "Sutrapur",
+      sub: "Dhaka Sadar HO",
+      postalCode: "1100",
+    },
+    {
+      city: "Sutrapur",
+      sub: "Wari TSO",
+      postalCode: "1203",
+    },
+    {
+      city: "Sutrapur",
+      sub: "Gendaria TSO",
+      postalCode: "1204",
+    },
+    {
+      city: "Tejgaon",
+      sub: "Tejgaon TSO",
+      postalCode: "1215",
+    },
+    {
+      city: "Tejgaon Industrial Area",
+      sub: "Dhaka Politechnic",
+      postalCode: "1208",
+    },
+    {
+      city: "Uttara",
+      sub: "Uttara Model TwonTSO",
+      postalCode: "1231",
+    },
+  ],
+  Mymensingh: [
+    {
+      city: "Bhaluka",
+      sub: "Bhaluka",
+      postalCode: "2240",
+    },
+    {
+      city: "Fulbaria",
+      sub: "Fulbaria",
+      postalCode: "2216",
+    },
+    {
+      city: "Gaforgaon",
+      sub: "Gaforgaon",
+      postalCode: "2230",
+    },
+    {
+      city: "Gaforgaon",
+      sub: "Shibganj",
+      postalCode: "2231",
+    },
+    {
+      city: "Gaforgaon",
+      sub: "Usti",
+      postalCode: "2232",
+    },
+    {
+      city: "Gaforgaon",
+      sub: "Kandipara",
+      postalCode: "2233",
+    },
+    {
+      city: "Gaforgaon",
+      sub: "Duttarbazar",
+      postalCode: "2234",
+    },
+    {
+      city: "Gouripur",
+      sub: "Gouripur",
+      postalCode: "2270",
+    },
+    {
+      city: "Gouripur",
+      sub: "Ramgopalpur",
+      postalCode: "2271",
+    },
+    {
+      city: "Haluaghat",
+      sub: "Haluaghat",
+      postalCode: "2260",
+    },
+    {
+      city: "Haluaghat",
+      sub: "Dhara",
+      postalCode: "2261",
+    },
+    {
+      city: "Haluaghat",
+      sub: "Munshirhat",
+      postalCode: "2262",
+    },
+    {
+      city: "Isshwargonj",
+      sub: "Isshwargonj",
+      postalCode: "2280",
+    },
+    {
+      city: "Isshwargonj",
+      sub: "Sohagi",
+      postalCode: "2281",
+    },
+    {
+      city: "Isshwargonj",
+      sub: "Atharabari",
+      postalCode: "2282",
+    },
+    {
+      city: "Muktagachha",
+      sub: "Muktagachha",
+      postalCode: "2210",
+    },
+    {
+      city: "Mymensingh Sadar",
+      sub: "Mymensingh Sadar",
+      postalCode: "2200",
+    },
+    {
+      city: "Mymensingh Sadar",
+      sub: "Kawatkhali",
+      postalCode: "2201",
+    },
+    {
+      city: "Mymensingh Sadar",
+      sub: "Agriculture Universi",
+      postalCode: "2202",
+    },
+    {
+      city: "Mymensingh Sadar",
+      sub: "Shombhuganj",
+      postalCode: "2203",
+    },
+    {
+      city: "Mymensingh Sadar",
+      sub: "Biddyaganj",
+      postalCode: "2204",
+    },
+    {
+      city: "Mymensingh Sadar",
+      sub: "Pearpur",
+      postalCode: "2205",
+    },
+    {
+      city: "Nandail",
+      sub: "Nandail",
+      postalCode: "2290",
+    },
+    {
+      city: "Nandail",
+      sub: "Gangail",
+      postalCode: "2291",
+    },
+    {
+      city: "Phulpur",
+      sub: "Phulpur",
+      postalCode: "2250",
+    },
+    {
+      city: "Phulpur",
+      sub: "Beltia",
+      postalCode: "2251",
+    },
+    {
+      city: "Phulpur",
+      sub: "Tarakanda",
+      postalCode: "2252",
+    },
+    {
+      city: "Trishal",
+      sub: "Trishal",
+      postalCode: "2220",
+    },
+    {
+      city: "Trishal",
+      sub: "Ahmadbad",
+      postalCode: "2221",
+    },
+    {
+      city: "Trishal",
+      sub: "Ram Amritaganj",
+      postalCode: "2222",
+    },
+    {
+      city: "Trishal",
+      sub: "Dhala",
+      postalCode: "2223",
+    },
+  ],
+  Kishoreganj: [
+    {
+      city: "Bajitpur",
+      sub: "Bajitpur",
+      postalCode: "2336",
+    },
+    {
+      city: "Bajitpur",
+      sub: "Sararchar",
+      postalCode: "2337",
+    },
+    {
+      city: "Bajitpur",
+      sub: "Laksmipur",
+      postalCode: "2338",
+    },
+    {
+      city: "Bhairob",
+      sub: "Bhairab",
+      postalCode: "2350",
+    },
+    {
+      city: "Hossenpur",
+      sub: "Hossenpur",
+      postalCode: "2320",
+    },
+    {
+      city: "Itna",
+      sub: "Itna",
+      postalCode: "2390",
+    },
+    {
+      city: "Karimganj",
+      sub: "Karimganj",
+      postalCode: "2310",
+    },
+    {
+      city: "Katiadi",
+      sub: "Katiadi",
+      postalCode: "2330",
+    },
+    {
+      city: "Katiadi",
+      sub: "Gochhihata",
+      postalCode: "2331",
+    },
+    {
+      city: "Kishoreganj Sadar",
+      sub: "Kishoreganj Sadar",
+      postalCode: "2300",
+    },
+    {
+      city: "Kishoreganj Sadar",
+      sub: "Kishoreganj S.Mills",
+      postalCode: "2301",
+    },
+    {
+      city: "Kishoreganj Sadar",
+      sub: "Maizhati",
+      postalCode: "2302",
+    },
+    {
+      city: "Kishoreganj Sadar",
+      sub: "Nilganj",
+      postalCode: "2303",
+    },
+    {
+      city: "Kuliarchar",
+      sub: "Kuliarchar",
+      postalCode: "2340",
+    },
+    {
+      city: "Kuliarchar",
+      sub: "Chhoysuti",
+      postalCode: "2341",
+    },
+    {
+      city: "Mithamoin",
+      sub: "MIthamoin",
+      postalCode: "2370",
+    },
+    {
+      city: "Mithamoin",
+      sub: "Abdullahpur",
+      postalCode: "2371",
+    },
+    {
+      city: "Nikli",
+      sub: "Nikli",
+      postalCode: "2360",
+    },
+    {
+      city: "Ostagram",
+      sub: "Ostagram",
+      postalCode: "2380",
+    },
+    {
+      city: "Pakundia",
+      sub: "Pakundia",
+      postalCode: "2326",
+    },
+    {
+      city: "Tarial",
+      sub: "Tarial",
+      postalCode: "2316",
+    },
+  ],
+  Netrakona: [
+    {
+      city: "Susung Durgapur",
+      sub: "IBH WAs Here",
+      postalCode: "$mart",
+    },
+    {
+      city: "Susung Durgapur",
+      sub: "Susnng Durgapur",
+      postalCode: "2420",
+    },
+    {
+      city: "Atpara",
+      sub: "Atpara",
+      postalCode: "2470",
+    },
+    {
+      city: "Barhatta",
+      sub: "Barhatta",
+      postalCode: "2440",
+    },
+    {
+      city: "Dharmapasha",
+      sub: "Dharampasha",
+      postalCode: "2450",
+    },
+    {
+      city: "Dhobaura",
+      sub: "Dhobaura",
+      postalCode: "2416",
+    },
+    {
+      city: "Dhobaura",
+      sub: "Sakoai",
+      postalCode: "2417",
+    },
+    {
+      city: "Kalmakanda",
+      sub: "Kalmakanda",
+      postalCode: "2430",
+    },
+    {
+      city: "Kendua",
+      sub: "Kendua",
+      postalCode: "2480",
+    },
+    {
+      city: "Khaliajuri",
+      sub: "Khaliajhri",
+      postalCode: "2460",
+    },
+    {
+      city: "Khaliajuri",
+      sub: "Shaldigha",
+      postalCode: "2462",
+    },
+    {
+      city: "Madan",
+      sub: "Madan",
+      postalCode: "2490",
+    },
+    {
+      city: "Moddhynagar",
+      sub: "Moddoynagar",
+      postalCode: "2456",
+    },
+    {
+      city: "Mohanganj",
+      sub: "Mohanganj",
+      postalCode: "2446",
+    },
+    {
+      city: "Netrakona Sadar",
+      sub: "Netrakona Sadar",
+      postalCode: "2400",
+    },
+    {
+      city: "Netrakona Sadar",
+      sub: "Baikherhati",
+      postalCode: "2401",
+    },
+    {
+      city: "Purbadhola",
+      sub: "Purbadhola",
+      postalCode: "2410",
+    },
+    {
+      city: "Purbadhola",
+      sub: "Shamgonj",
+      postalCode: "2411",
+    },
+    {
+      city: "Purbadhola",
+      sub: "Jaria Jhanjhail",
+      postalCode: "2412",
+    },
+  ],
+  Narayanganj: [
+    {
+      city: "Araihazar",
+      sub: "Araihazar",
+      postalCode: "1450",
+    },
+    {
+      city: "Araihazar",
+      sub: "Gopaldi",
+      postalCode: "1451",
+    },
+    {
+      city: "Baidder Bazar",
+      sub: "Baidder Bazar",
+      postalCode: "1440",
+    },
+    {
+      city: "Baidder Bazar",
+      sub: "Bara Nagar",
+      postalCode: "1441",
+    },
+    {
+      city: "Baidder Bazar",
+      sub: "Barodi",
+      postalCode: "1442",
+    },
+    {
+      city: "Bandar",
+      sub: "Bandar",
+      postalCode: "1410",
+    },
+    {
+      city: "Bandar",
+      sub: "D.C Mills",
+      postalCode: "1411",
+    },
+    {
+      city: "Bandar",
+      sub: "Nabiganj",
+      postalCode: "1412",
+    },
+    {
+      city: "Bandar",
+      sub: "BIDS",
+      postalCode: "1413",
+    },
+    {
+      city: "Bandar",
+      sub: "Madanganj",
+      postalCode: "1414",
+    },
+    {
+      city: "Fatullah",
+      sub: "Fatullah",
+      postalCode: "1420",
+    },
+    {
+      city: "Fatullah",
+      sub: "Fatulla Bazar",
+      postalCode: "1421",
+    },
+    {
+      city: "Narayanganj Sadar",
+      sub: "Narayanganj Sadar",
+      postalCode: "1400",
+    },
+    {
+      city: "Rupganj",
+      sub: "Rupganj",
+      postalCode: "1460",
+    },
+    {
+      city: "Rupganj",
+      sub: "Kanchan",
+      postalCode: "1461",
+    },
+    {
+      city: "Rupganj",
+      sub: "Bhulta",
+      postalCode: "1462",
+    },
+    {
+      city: "Rupganj",
+      sub: "Nagri",
+      postalCode: "1463",
+    },
+    {
+      city: "Rupganj",
+      sub: "Murapara",
+      postalCode: "1464",
+    },
+    {
+      city: "Siddirganj",
+      sub: "Siddirganj",
+      postalCode: "1430",
+    },
+    {
+      city: "Siddirganj",
+      sub: "Adamjeenagar",
+      postalCode: "1431",
+    },
+    {
+      city: "Siddirganj",
+      sub: "LN Mills",
+      postalCode: "1432",
+    },
+  ],
+  Munshiganj: [
+    {
+      city: "Gajaria",
+      sub: "Gajaria",
+      postalCode: "1510",
+    },
+    {
+      city: "Gajaria",
+      sub: "Hossendi",
+      postalCode: "1511",
+    },
+    {
+      city: "Gajaria",
+      sub: "Rasulpur",
+      postalCode: "1512",
+    },
+    {
+      city: "Lohajong",
+      sub: "Haridia",
+      postalCode: "1333",
+    },
+    {
+      city: "Lohajong",
+      sub: "Gouragonj",
+      postalCode: "1334",
+    },
+    {
+      city: "Lohajong",
+      sub: "Madini Mandal",
+      postalCode: "1335",
+    },
+    {
+      city: "Lohajong",
+      sub: "Lohajang",
+      postalCode: "1530",
+    },
+    {
+      city: "Lohajong",
+      sub: "Korhati",
+      postalCode: "1531",
+    },
+    {
+      city: "Lohajong",
+      sub: "Haldia SO",
+      postalCode: "1532",
+    },
+    {
+      city: "Lohajong",
+      sub: "Haridia DESO",
+      postalCode: "1533",
+    },
+    {
+      city: "Lohajong",
+      sub: "Gouragonj",
+      postalCode: "1534",
+    },
+    {
+      city: "Lohajong",
+      sub: "Medini Mandal EDSO",
+      postalCode: "1535",
+    },
+    {
+      city: "Munshiganj Sadar",
+      sub: "Munshiganj Sadar",
+      postalCode: "1500",
+    },
+    {
+      city: "Munshiganj Sadar",
+      sub: "Rikabibazar",
+      postalCode: "1501",
+    },
+    {
+      city: "Munshiganj Sadar",
+      sub: "Mirkadim",
+      postalCode: "1502",
+    },
+    {
+      city: "Munshiganj Sadar",
+      sub: "Kathakhali",
+      postalCode: "1503",
+    },
+    {
+      city: "Sirajdikhan",
+      sub: "Sirajdikhan",
+      postalCode: "1540",
+    },
+    {
+      city: "Sirajdikhan",
+      sub: "Kola",
+      postalCode: "1541",
+    },
+    {
+      city: "Sirajdikhan",
+      sub: "Ichapur",
+      postalCode: "1542",
+    },
+    {
+      city: "Sirajdikhan",
+      sub: "Malkha Nagar",
+      postalCode: "1543",
+    },
+    {
+      city: "Sirajdikhan",
+      sub: "Shekher Nagar",
+      postalCode: "1544",
+    },
+    {
+      city: "Srinagar",
+      sub: "Srinagar",
+      postalCode: "1550",
+    },
+    {
+      city: "Srinagar",
+      sub: "Barikhal",
+      postalCode: "1551",
+    },
+    {
+      city: "Srinagar",
+      sub: "Mazpara",
+      postalCode: "1552",
+    },
+    {
+      city: "Srinagar",
+      sub: "Hashara",
+      postalCode: "1553",
+    },
+    {
+      city: "Srinagar",
+      sub: "Kolapara",
+      postalCode: "1554",
+    },
+    {
+      city: "Srinagar",
+      sub: "Kumarbhog",
+      postalCode: "1555",
+    },
+    {
+      city: "Srinagar",
+      sub: "Vaggyakul SO",
+      postalCode: "1556",
+    },
+    {
+      city: "Srinagar",
+      sub: "Baghra",
+      postalCode: "1557",
+    },
+    {
+      city: "Srinagar",
+      sub: "Bhaggyakul",
+      postalCode: "1558",
+    },
+    {
+      city: "Tangibari",
+      sub: "Tangibari",
+      postalCode: "1520",
+    },
+    {
+      city: "Tangibari",
+      sub: "Betkahat",
+      postalCode: "1521",
+    },
+    {
+      city: "Tangibari",
+      sub: "Baligao",
+      postalCode: "1522",
+    },
+    {
+      city: "Tangibari",
+      sub: "Bajrajugini",
+      postalCode: "1523",
+    },
+    {
+      city: "Tangibari",
+      sub: "Hasail",
+      postalCode: "1524",
+    },
+    {
+      city: "Tangibari",
+      sub: "Dighirpar",
+      postalCode: "1525",
+    },
+    {
+      city: "Tangibari",
+      sub: "Pura EDSO",
+      postalCode: "1526",
+    },
+    {
+      city: "Tangibari",
+      sub: "Pura",
+      postalCode: "1527",
+    },
+  ],
+  Narshingdi: [
+    {
+      city: "Belabo",
+      sub: "Belabo",
+      postalCode: "1640",
+    },
+    {
+      city: "Monohordi",
+      sub: "Monohordi",
+      postalCode: "1650",
+    },
+    {
+      city: "Monohordi",
+      sub: "Hatirdia",
+      postalCode: "1651",
+    },
+    {
+      city: "Monohordi",
+      sub: "Katabaria",
+      postalCode: "1652",
+    },
+    {
+      city: "Narshingdi Sadar",
+      sub: "Narshingdi Sadar",
+      postalCode: "1600",
+    },
+    {
+      city: "Narshingdi Sadar",
+      sub: "UMC Jute Mills",
+      postalCode: "1601",
+    },
+    {
+      city: "Narshingdi Sadar",
+      sub: "Narshingdi College",
+      postalCode: "1602",
+    },
+    {
+      city: "Narshingdi Sadar",
+      sub: "Panchdona",
+      postalCode: "1603",
+    },
+    {
+      city: "Narshingdi Sadar",
+      sub: "Madhabdi",
+      postalCode: "1604",
+    },
+    {
+      city: "Narshingdi Sadar",
+      sub: "Karimpur",
+      postalCode: "1605",
+    },
+    {
+      city: "Palash",
+      sub: "Palash",
+      postalCode: "1610",
+    },
+    {
+      city: "Palash",
+      sub: "Ghorashal Urea Facto",
+      postalCode: "1611",
+    },
+    {
+      city: "Palash",
+      sub: "Char Sindhur",
+      postalCode: "1612",
+    },
+    {
+      city: "Palash",
+      sub: "Ghorashal",
+      postalCode: "1613",
+    },
+    {
+      city: "Raypura",
+      sub: "Raypura",
+      postalCode: "1630",
+    },
+    {
+      city: "Raypura",
+      sub: "Bazar Hasnabad",
+      postalCode: "1631",
+    },
+    {
+      city: "Raypura",
+      sub: "Radhaganj bazar",
+      postalCode: "1632",
+    },
+    {
+      city: "Shibpur",
+      sub: "Shibpur",
+      postalCode: "1620",
+    },
+  ],
+  Gazipur: [
+    {
+      city: "Gazipur Sadar",
+      sub: "Gazipur Sadar",
+      postalCode: "1700",
+    },
+    {
+      city: "Gazipur Sadar",
+      sub: "B.R.R",
+      postalCode: "1701",
+    },
+    {
+      city: "Gazipur Sadar",
+      sub: "Chandna",
+      postalCode: "1702",
+    },
+    {
+      city: "Gazipur Sadar",
+      sub: "B.O.F",
+      postalCode: "1703",
+    },
+    {
+      city: "Gazipur Sadar",
+      sub: "National University",
+      postalCode: "1704",
+    },
+    {
+      city: "Kaliakaar",
+      sub: "Kaliakaar",
+      postalCode: "1750",
+    },
+    {
+      city: "Kaliakaar",
+      sub: "Safipur",
+      postalCode: "1751",
+    },
+    {
+      city: "Kaliganj",
+      sub: "Kaliganj",
+      postalCode: "1720",
+    },
+    {
+      city: "Kaliganj",
+      sub: "Pubail",
+      postalCode: "1721",
+    },
+    {
+      city: "Kaliganj",
+      sub: "Santanpara",
+      postalCode: "1722",
+    },
+    {
+      city: "Kaliganj",
+      sub: "Vaoal Jamalpur",
+      postalCode: "1723",
+    },
+    {
+      city: "Kapashia",
+      sub: "kapashia",
+      postalCode: "1730",
+    },
+    {
+      city: "Monnunagar",
+      sub: "Monnunagar",
+      postalCode: "1710",
+    },
+    {
+      city: "Monnunagar",
+      sub: "Nishat Nagar",
+      postalCode: "1711",
+    },
+    {
+      city: "Monnunagar",
+      sub: "Ershad Nagar",
+      postalCode: "1712",
+    },
+    {
+      city: "Sreepur",
+      sub: "Sreepur",
+      postalCode: "1740",
+    },
+    {
+      city: "Sreepur",
+      sub: "Barmi",
+      postalCode: "1743",
+    },
+    {
+      city: "Sreepur",
+      sub: "Satkhamair",
+      postalCode: "1744",
+    },
+    {
+      city: "Sreepur",
+      sub: "Kawraid",
+      postalCode: "1745",
+    },
+    {
+      city: "Sreepur",
+      sub: "Bashamur",
+      postalCode: "1747",
+    },
+    {
+      city: "Sreepur",
+      sub: "Boubi",
+      postalCode: "1748",
+    },
+    {
+      city: "Sripur",
+      sub: "Rajendrapur",
+      postalCode: "1741",
+    },
+    {
+      city: "Sripur",
+      sub: "Rajendrapur Canttome",
+      postalCode: "1742",
+    },
+  ],
+  Rajbari: [
+    {
+      city: "Baliakandi",
+      sub: "Baliakandi",
+      postalCode: "7730",
+    },
+    {
+      city: "Baliakandi",
+      sub: "Nalia",
+      postalCode: "7731",
+    },
+    {
+      city: "Pangsha",
+      sub: "Pangsha",
+      postalCode: "7720",
+    },
+    {
+      city: "Pangsha",
+      sub: "Ramkol",
+      postalCode: "7721",
+    },
+    {
+      city: "Pangsha",
+      sub: "Ratandia",
+      postalCode: "7722",
+    },
+    {
+      city: "Pangsha",
+      sub: "Mrigibazar",
+      postalCode: "7723",
+    },
+    {
+      city: "Rajbari Sadar",
+      sub: "Rajbari Sadar",
+      postalCode: "7700",
+    },
+    {
+      city: "Rajbari Sadar",
+      sub: "Goalanda",
+      postalCode: "7710",
+    },
+    {
+      city: "Rajbari Sadar",
+      sub: "Khankhanapur",
+      postalCode: "7711",
+    },
+  ],
+  Faridpur: [
+    {
+      city: "Alfadanga",
+      sub: "Alfadanga",
+      postalCode: "7870",
+    },
+    {
+      city: "Bhanga",
+      sub: "Bhanga",
+      postalCode: "7830",
+    },
+    {
+      city: "Boalmari",
+      sub: "Boalmari",
+      postalCode: "7860",
+    },
+    {
+      city: "Boalmari",
+      sub: "Rupatpat",
+      postalCode: "7861",
+    },
+    {
+      city: "Charbhadrasan",
+      sub: "Charbadrashan",
+      postalCode: "7810",
+    },
+    {
+      city: "Faridpur Sadar",
+      sub: "Faridpursadar",
+      postalCode: "7800",
+    },
+    {
+      city: "Faridpur Sadar",
+      sub: "Kanaipur",
+      postalCode: "7801",
+    },
+    {
+      city: "Faridpur Sadar",
+      sub: "Ambikapur",
+      postalCode: "7802",
+    },
+    {
+      city: "Faridpur Sadar",
+      sub: "Baitulaman Politecni",
+      postalCode: "7803",
+    },
+    {
+      city: "Madukhali",
+      sub: "Madukhali",
+      postalCode: "7850",
+    },
+    {
+      city: "Madukhali",
+      sub: "Kamarkali",
+      postalCode: "7851",
+    },
+    {
+      city: "Nagarkanda",
+      sub: "Nagarkanda",
+      postalCode: "7840",
+    },
+    {
+      city: "Nagarkanda",
+      sub: "Talma",
+      postalCode: "7841",
+    },
+    {
+      city: "Sadarpur",
+      sub: "Sadarpur",
+      postalCode: "7820",
+    },
+    {
+      city: "Sadarpur",
+      sub: "Hat Krishapur",
+      postalCode: "7821",
+    },
+    {
+      city: "Sadarpur",
+      sub: "Bishwa jaker Manjil",
+      postalCode: "7822",
+    },
+    {
+      city: "Shriangan",
+      sub: "Shriangan",
+      postalCode: "7804",
+    },
+  ],
+  Madaripur: [
+    {
+      city: "Barhamganj",
+      sub: "Barhamganj",
+      postalCode: "7930",
+    },
+    {
+      city: "Barhamganj",
+      sub: "Nilaksmibandar",
+      postalCode: "7931",
+    },
+    {
+      city: "Barhamganj",
+      sub: "Bahadurpur",
+      postalCode: "7932",
+    },
+    {
+      city: "Barhamganj",
+      sub: "Umedpur",
+      postalCode: "7933",
+    },
+    {
+      city: "kalkini",
+      sub: "Kalkini",
+      postalCode: "7920",
+    },
+    {
+      city: "kalkini",
+      sub: "Sahabrampur",
+      postalCode: "7921",
+    },
+    {
+      city: "Madaripur Sadar",
+      sub: "Madaripur Sadar",
+      postalCode: "7900",
+    },
+    {
+      city: "Madaripur Sadar",
+      sub: "Charmugria",
+      postalCode: "7901",
+    },
+    {
+      city: "Madaripur Sadar",
+      sub: "Kulpaddi",
+      postalCode: "7902",
+    },
+    {
+      city: "Madaripur Sadar",
+      sub: "Habiganj",
+      postalCode: "7903",
+    },
+    {
+      city: "Madaripur Sadar",
+      sub: "Mustafapur",
+      postalCode: "7904",
+    },
+    {
+      city: "Rajoir",
+      sub: "Rajoir",
+      postalCode: "7910",
+    },
+    {
+      city: "Rajoir",
+      sub: "Khalia",
+      postalCode: "7911",
+    },
+  ],
+  Shariatpur: [
+    {
+      city: "Bhedorganj",
+      sub: "Bhedorganj",
+      postalCode: "8030",
+    },
+    {
+      city: "Damudhya",
+      sub: "Damudhya",
+      postalCode: "8040",
+    },
+    {
+      city: "Gosairhat",
+      sub: "Gosairhat",
+      postalCode: "8050",
+    },
+    {
+      city: "Jajira",
+      sub: "Jajira",
+      postalCode: "8010",
+    },
+    {
+      city: "Naria",
+      sub: "Naria",
+      postalCode: "8020",
+    },
+    {
+      city: "Naria",
+      sub: "Bhozeshwar",
+      postalCode: "8021",
+    },
+    {
+      city: "Naria",
+      sub: "Gharisar",
+      postalCode: "8022",
+    },
+    {
+      city: "Naria",
+      sub: "Upshi",
+      postalCode: "8023",
+    },
+    {
+      city: "Naria",
+      sub: "Kartikpur",
+      postalCode: "8024",
+    },
+    {
+      city: "Shariatpur Sadar",
+      sub: "Shariatpur Sadar",
+      postalCode: "8000",
+    },
+    {
+      city: "Shariatpur Sadar",
+      sub: "Angaria",
+      postalCode: "8001",
+    },
+    {
+      city: "Shariatpur Sadar",
+      sub: "Chikandi",
+      postalCode: "8002",
+    },
+  ],
+  Gopalganj: [
+    {
+      city: "Gopalganj Sadar",
+      sub: "Chandradighalia",
+      postalCode: "8013",
+    },
+    {
+      city: "Gopalganj Sadar",
+      sub: "Gopalganj Sadar",
+      postalCode: "8100",
+    },
+    {
+      city: "Gopalganj Sadar",
+      sub: "Ulpur",
+      postalCode: "8101",
+    },
+    {
+      city: "Gopalganj Sadar",
+      sub: "Barfa",
+      postalCode: "8102",
+    },
+    {
+      city: "Kashiani",
+      sub: "Kashiani",
+      postalCode: "8130",
+    },
+    {
+      city: "Kashiani",
+      sub: "Ramdia College",
+      postalCode: "8131",
+    },
+    {
+      city: "Kashiani",
+      sub: "Ratoil",
+      postalCode: "8132",
+    },
+    {
+      city: "Kashiani",
+      sub: "Jonapur",
+      postalCode: "8133",
+    },
+    {
+      city: "Kotalipara",
+      sub: "Kotalipara",
+      postalCode: "8110",
+    },
+    {
+      city: "Maksudpur",
+      sub: "Maksudpur",
+      postalCode: "8140",
+    },
+    {
+      city: "Maksudpur",
+      sub: "Batkiamari",
+      postalCode: "8141",
+    },
+    {
+      city: "Maksudpur",
+      sub: "Khandarpara",
+      postalCode: "8142",
+    },
+    {
+      city: "Tungipara",
+      sub: "Tungipara",
+      postalCode: "8120",
+    },
+    {
+      city: "Tungipara",
+      sub: "Patgati",
+      postalCode: "8121",
+    },
+  ],
+  Manikganj: [
+    {
+      city: "Doulatpur",
+      sub: "Doulatpur",
+      postalCode: "1860",
+    },
+    {
+      city: "Gheor",
+      sub: "Gheor",
+      postalCode: "1840",
+    },
+    {
+      city: "Lechhraganj",
+      sub: "Lechhraganj",
+      postalCode: "1830",
+    },
+    {
+      city: "Lechhraganj",
+      sub: "Jhitka",
+      postalCode: "1831",
+    },
+    {
+      city: "Manikganj Sadar",
+      sub: "Manikganj Sadar",
+      postalCode: "1800",
+    },
+    {
+      city: "Manikganj Sadar",
+      sub: "Manikganj Bazar",
+      postalCode: "1801",
+    },
+    {
+      city: "Manikganj Sadar",
+      sub: "Gorpara",
+      postalCode: "1802",
+    },
+    {
+      city: "Manikganj Sadar",
+      sub: "Mahadebpur",
+      postalCode: "1803",
+    },
+    {
+      city: "Manikganj Sadar",
+      sub: "Barangail",
+      postalCode: "1804",
+    },
+    {
+      city: "Saturia",
+      sub: "Saturia",
+      postalCode: "1810",
+    },
+    {
+      city: "Saturia",
+      sub: "Baliati",
+      postalCode: "1811",
+    },
+    {
+      city: "Shibloya",
+      sub: "Shibaloy",
+      postalCode: "1850",
+    },
+    {
+      city: "Shibloya",
+      sub: "Aricha",
+      postalCode: "1851",
+    },
+    {
+      city: "Shibloya",
+      sub: "Tewta",
+      postalCode: "1852",
+    },
+    {
+      city: "Shibloya",
+      sub: "Uthli",
+      postalCode: "1853",
+    },
+    {
+      city: "Singari",
+      sub: "Singair",
+      postalCode: "1820",
+    },
+    {
+      city: "Singari",
+      sub: "Baira",
+      postalCode: "1821",
+    },
+    {
+      city: "Singari",
+      sub: "joymantop",
+      postalCode: "1822",
+    },
+  ],
+  Tangail: [
+    {
+      city: "Basail",
+      sub: "Basail",
+      postalCode: "1920",
+    },
+    {
+      city: "Bhuapur",
+      sub: "Bhuapur",
+      postalCode: "1960",
+    },
+    {
+      city: "Delduar",
+      sub: "Delduar",
+      postalCode: "1910",
+    },
+    {
+      city: "Delduar",
+      sub: "Jangalia",
+      postalCode: "1911",
+    },
+    {
+      city: "Delduar",
+      sub: "Patharail",
+      postalCode: "1912",
+    },
+    {
+      city: "Delduar",
+      sub: "Elasin",
+      postalCode: "1913",
+    },
+    {
+      city: "Delduar",
+      sub: "Hinga Nagar",
+      postalCode: "1914",
+    },
+    {
+      city: "Delduar",
+      sub: "Lowhati",
+      postalCode: "1915",
+    },
+    {
+      city: "Ghatail",
+      sub: "Ghatial",
+      postalCode: "1980",
+    },
+    {
+      city: "Ghatail",
+      sub: "Zahidganj",
+      postalCode: "1981",
+    },
+    {
+      city: "Ghatail",
+      sub: "D. Pakutia",
+      postalCode: "1982",
+    },
+    {
+      city: "Ghatail",
+      sub: "Dhalapara",
+      postalCode: "1983",
+    },
+    {
+      city: "Ghatail",
+      sub: "Lohani",
+      postalCode: "1984",
+    },
+    {
+      city: "Gopalpur",
+      sub: "Gopalpur",
+      postalCode: "1990",
+    },
+    {
+      city: "Gopalpur",
+      sub: "Jhowail",
+      postalCode: "1991",
+    },
+    {
+      city: "Gopalpur",
+      sub: "Hemnagar",
+      postalCode: "1992",
+    },
+    {
+      city: "Kalihati",
+      sub: "Kalihati",
+      postalCode: "1970",
+    },
+    {
+      city: "Kalihati",
+      sub: "Rajafair",
+      postalCode: "1971",
+    },
+    {
+      city: "Kalihati",
+      sub: "Nagbari",
+      postalCode: "1972",
+    },
+    {
+      city: "Kalihati",
+      sub: "Ballabazar",
+      postalCode: "1973",
+    },
+    {
+      city: "Kalihati",
+      sub: "Elinga",
+      postalCode: "1974",
+    },
+    {
+      city: "Kalihati",
+      sub: "Palisha",
+      postalCode: "1975",
+    },
+    {
+      city: "Kalihati",
+      sub: "Nagarbari SO",
+      postalCode: "1976",
+    },
+    {
+      city: "Kalihati",
+      sub: "Nagarbari",
+      postalCode: "1977",
+    },
+    {
+      city: "Kashkaolia",
+      sub: "Kashkawlia",
+      postalCode: "1930",
+    },
+    {
+      city: "Madhupur",
+      sub: "Madhupur",
+      postalCode: "1996",
+    },
+    {
+      city: "Madhupur",
+      sub: "Dhobari",
+      postalCode: "1997",
+    },
+    {
+      city: "Mirzapur",
+      sub: "Mirzapur",
+      postalCode: "1940",
+    },
+    {
+      city: "Mirzapur",
+      sub: "Gorai",
+      postalCode: "1941",
+    },
+    {
+      city: "Mirzapur",
+      sub: "M.C. College",
+      postalCode: "1942",
+    },
+    {
+      city: "Mirzapur",
+      sub: "Warri paikpara",
+      postalCode: "1943",
+    },
+    {
+      city: "Mirzapur",
+      sub: "Jarmuki",
+      postalCode: "1944",
+    },
+    {
+      city: "Mirzapur",
+      sub: "Mohera",
+      postalCode: "1945",
+    },
+    {
+      city: "Nagarpur",
+      sub: "Nagarpur",
+      postalCode: "1936",
+    },
+    {
+      city: "Nagarpur",
+      sub: "Dhuburia",
+      postalCode: "1937",
+    },
+    {
+      city: "Nagarpur",
+      sub: "Salimabad",
+      postalCode: "1938",
+    },
+    {
+      city: "Sakhipur",
+      sub: "Sakhipur",
+      postalCode: "1950",
+    },
+    {
+      city: "Sakhipur",
+      sub: "Kochua",
+      postalCode: "1951",
+    },
+    {
+      city: "Tangail Sadar",
+      sub: "Tangail Sadar",
+      postalCode: "1900",
+    },
+    {
+      city: "Tangail Sadar",
+      sub: "Kagmari",
+      postalCode: "1901",
+    },
+    {
+      city: "Tangail Sadar",
+      sub: "Santosh",
+      postalCode: "1902",
+    },
+    {
+      city: "Tangail Sadar",
+      sub: "Korotia",
+      postalCode: "1903",
+    },
+    {
+      city: "Tangail Sadar",
+      sub: "Purabari",
+      postalCode: "1904",
+    },
+  ],
+  Jamalpur: [
+    {
+      city: "Dewangonj",
+      sub: "Dewangonj",
+      postalCode: "2030",
+    },
+    {
+      city: "Dewangonj",
+      sub: "Dewangonj S. Mills",
+      postalCode: "2031",
+    },
+    {
+      city: "Islampur",
+      sub: "Islampur",
+      postalCode: "2020",
+    },
+    {
+      city: "Islampur",
+      sub: "Durmoot",
+      postalCode: "2021",
+    },
+    {
+      city: "Islampur",
+      sub: "Gilabari",
+      postalCode: "2022",
+    },
+    {
+      city: "Jamalpur",
+      sub: "Jamalpur",
+      postalCode: "2000",
+    },
+    {
+      city: "Jamalpur",
+      sub: "Nandina",
+      postalCode: "2001",
+    },
+    {
+      city: "Jamalpur",
+      sub: "Narundi",
+      postalCode: "2002",
+    },
+    {
+      city: "Malandah",
+      sub: "Malandah",
+      postalCode: "2010",
+    },
+    {
+      city: "Malandah",
+      sub: "Jamalpur",
+      postalCode: "2011",
+    },
+    {
+      city: "Malandah",
+      sub: "Malancha",
+      postalCode: "2012",
+    },
+    {
+      city: "Malandah",
+      sub: "Mahmoodpur",
+      postalCode: "2013",
+    },
+    {
+      city: "Mathargonj",
+      sub: "Mathargonj",
+      postalCode: "2040",
+    },
+    {
+      city: "Mathargonj",
+      sub: "Balijhuri",
+      postalCode: "2041",
+    },
+    {
+      city: "Shorishabari",
+      sub: "Shorishabari",
+      postalCode: "2050",
+    },
+    {
+      city: "Shorishabari",
+      sub: "Gunerbari",
+      postalCode: "2051",
+    },
+    {
+      city: "Shorishabari",
+      sub: "Bausee",
+      postalCode: "2052",
+    },
+    {
+      city: "Shorishabari",
+      sub: "Jagannath Ghat",
+      postalCode: "2053",
+    },
+    {
+      city: "Shorishabari",
+      sub: "Pingna",
+      postalCode: "2054",
+    },
+    {
+      city: "Shorishabari",
+      sub: "Jamuna Sar Karkhana",
+      postalCode: "2055",
+    },
+  ],
+  Sherpur: [
+    {
+      city: "Bakshigonj",
+      sub: "Bakshigonj",
+      postalCode: "2140",
+    },
+    {
+      city: "Jhinaigati",
+      sub: "Jhinaigati",
+      postalCode: "2120",
+    },
+    {
+      city: "Nakla",
+      sub: "Nakla",
+      postalCode: "2150",
+    },
+    {
+      city: "Nakla",
+      sub: "Gonopaddi",
+      postalCode: "2151",
+    },
+    {
+      city: "Nalitabari",
+      sub: "Nalitabari",
+      postalCode: "2110",
+    },
+    {
+      city: "Nalitabari",
+      sub: "Hatibandha",
+      postalCode: "2111",
+    },
+    {
+      city: "Sherpur Shadar",
+      sub: "Sherpur Shadar",
+      postalCode: "2100",
+    },
+    {
+      city: "Shribardi",
+      sub: "Shribardi",
+      postalCode: "2130",
+    },
+  ],
+  Barishal: [
+    {
+      city: "Agailzhara",
+      sub: "Agailzhara",
+      postalCode: "8240",
+    },
+    {
+      city: "Agailzhara",
+      sub: "Gaila",
+      postalCode: "8241",
+    },
+    {
+      city: "Agailzhara",
+      sub: "Paisarhat",
+      postalCode: "8242",
+    },
+    {
+      city: "Babuganj",
+      sub: "Babuganj",
+      postalCode: "8210",
+    },
+    {
+      city: "Babuganj",
+      sub: "Rahamatpur",
+      postalCode: "8211",
+    },
+    {
+      city: "Babuganj",
+      sub: "Chandpasha",
+      postalCode: "8212",
+    },
+    {
+      city: "Babuganj",
+      sub: "Madhabpasha",
+      postalCode: "8213",
+    },
+    {
+      city: "Babuganj",
+      sub: "Thakur Mallik",
+      postalCode: "8214",
+    },
+    {
+      city: "Babuganj",
+      sub: "Nizamuddin College",
+      postalCode: "8215",
+    },
+    {
+      city: "Babuganj",
+      sub: "Barishal Cadet",
+      postalCode: "8216",
+    },
+    {
+      city: "Barajalia",
+      sub: "Barajalia",
+      postalCode: "8260",
+    },
+    {
+      city: "Barajalia",
+      sub: "Osman Manjil",
+      postalCode: "8261",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Barishal Sadar",
+      postalCode: "8200",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Bukhainagar",
+      postalCode: "8201",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Saheberhat",
+      postalCode: "8202",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Sugandia",
+      postalCode: "8203",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Patang",
+      postalCode: "8204",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Kashipur",
+      postalCode: "8205",
+    },
+    {
+      city: "Barishal Sadar",
+      sub: "Jaguarhat",
+      postalCode: "8206",
+    },
+    {
+      city: "Gouranadi",
+      sub: "Gouranadi",
+      postalCode: "8230",
+    },
+    {
+      city: "Gouranadi",
+      sub: "Tarki Bandar",
+      postalCode: "8231",
+    },
+    {
+      city: "Gouranadi",
+      sub: "Kashemabad",
+      postalCode: "8232",
+    },
+    {
+      city: "Gouranadi",
+      sub: "Batajor",
+      postalCode: "8233",
+    },
+    {
+      city: "Mahendiganj",
+      sub: "Mahendiganj",
+      postalCode: "8270",
+    },
+    {
+      city: "Mahendiganj",
+      sub: "Laskarpur",
+      postalCode: "8271",
+    },
+    {
+      city: "Mahendiganj",
+      sub: "Ulania",
+      postalCode: "8272",
+    },
+    {
+      city: "Mahendiganj",
+      sub: "Nalgora",
+      postalCode: "8273",
+    },
+    {
+      city: "Mahendiganj",
+      sub: "Langutia",
+      postalCode: "8274",
+    },
+    {
+      city: "Muladi",
+      sub: "Muladi",
+      postalCode: "8250",
+    },
+    {
+      city: "Muladi",
+      sub: "Kazirchar",
+      postalCode: "8251",
+    },
+    {
+      city: "Muladi",
+      sub: "Charkalekhan",
+      postalCode: "8252",
+    },
+    {
+      city: "Sahebganj",
+      sub: "Sahebganj",
+      postalCode: "8280",
+    },
+    {
+      city: "Sahebganj",
+      sub: "Charamandi",
+      postalCode: "8281",
+    },
+    {
+      city: "Sahebganj",
+      sub: "Padri Shibpur",
+      postalCode: "8282",
+    },
+    {
+      city: "Sahebganj",
+      sub: "Shialguni",
+      postalCode: "8283",
+    },
+    {
+      city: "Sahebganj",
+      sub: "kalaskati",
+      postalCode: "8284",
+    },
+    {
+      city: "Uzirpur",
+      sub: "Uzirpur",
+      postalCode: "8220",
+    },
+    {
+      city: "Uzirpur",
+      sub: "Dhamura",
+      postalCode: "8221",
+    },
+    {
+      city: "Uzirpur",
+      sub: "Jugirkanda",
+      postalCode: "8222",
+    },
+    {
+      city: "Uzirpur",
+      sub: "Dakuarhat",
+      postalCode: "8223",
+    },
+    {
+      city: "Uzirpur",
+      sub: "Shikarpur",
+      postalCode: "8224",
+    },
+  ],
+  Bhola: [
+    {
+      city: "Bhola Sadar",
+      sub: "Bhola Sadar",
+      postalCode: "8300",
+    },
+    {
+      city: "Bhola Sadar",
+      sub: "Joynagar",
+      postalCode: "8301",
+    },
+    {
+      city: "Borhanuddin UPO",
+      sub: "Borhanuddin UPO",
+      postalCode: "8320",
+    },
+    {
+      city: "Borhanuddin UPO",
+      sub: "Mirzakalu",
+      postalCode: "8321",
+    },
+    {
+      city: "Charfashion",
+      sub: "Charfashion",
+      postalCode: "8340",
+    },
+    {
+      city: "Charfashion",
+      sub: "Dularhat",
+      postalCode: "8341",
+    },
+    {
+      city: "Charfashion",
+      sub: "Keramatganj",
+      postalCode: "8342",
+    },
+    {
+      city: "Doulatkhan",
+      sub: "Doulatkhan",
+      postalCode: "8310",
+    },
+    {
+      city: "Doulatkhan",
+      sub: "Hajipur",
+      postalCode: "8311",
+    },
+    {
+      city: "Hajirhat",
+      sub: "Hajirhat",
+      postalCode: "8360",
+    },
+    {
+      city: "Hatshoshiganj",
+      sub: "Hatshoshiganj",
+      postalCode: "8350",
+    },
+    {
+      city: "Lalmohan UPO",
+      sub: "Lalmohan UPO",
+      postalCode: "8330",
+    },
+    {
+      city: "Lalmohan UPO",
+      sub: "Daurihat",
+      postalCode: "8331",
+    },
+    {
+      city: "Lalmohan UPO",
+      sub: "Gazaria",
+      postalCode: "8332",
+    },
+  ],
+  Jhalokathi: [
+    {
+      city: "Jhalokathi Sadar",
+      sub: "Jhalokathi Sadar",
+      postalCode: "8400",
+    },
+    {
+      city: "Jhalokathi Sadar",
+      sub: "Nabagram",
+      postalCode: "8401",
+    },
+    {
+      city: "Jhalokathi Sadar",
+      sub: "Baukathi",
+      postalCode: "8402",
+    },
+    {
+      city: "Jhalokathi Sadar",
+      sub: "Gabha",
+      postalCode: "8403",
+    },
+    {
+      city: "Jhalokathi Sadar",
+      sub: "Shekherhat",
+      postalCode: "8404",
+    },
+    {
+      city: "Kathalia",
+      sub: "Kathalia",
+      postalCode: "8430",
+    },
+    {
+      city: "Kathalia",
+      sub: "Amua",
+      postalCode: "8431",
+    },
+    {
+      city: "Kathalia",
+      sub: "Niamatee",
+      postalCode: "8432",
+    },
+    {
+      city: "Kathalia",
+      sub: "Shoulajalia",
+      postalCode: "8433",
+    },
+    {
+      city: "Nalchhiti",
+      sub: "Nalchhiti",
+      postalCode: "8420",
+    },
+    {
+      city: "Nalchhiti",
+      sub: "Beerkathi",
+      postalCode: "8421",
+    },
+    {
+      city: "Rajapur",
+      sub: "Rajapur",
+      postalCode: "8410",
+    },
+  ],
+  Pirojpur: [
+    {
+      city: "Banaripara",
+      sub: "Banaripara",
+      postalCode: "8530",
+    },
+    {
+      city: "Banaripara",
+      sub: "Chakhar",
+      postalCode: "8531",
+    },
+    {
+      city: "Bhandaria",
+      sub: "Bhandaria",
+      postalCode: "8550",
+    },
+    {
+      city: "Bhandaria",
+      sub: "Kanudashkathi",
+      postalCode: "8551",
+    },
+    {
+      city: "Bhandaria",
+      sub: "Dhaoa",
+      postalCode: "8552",
+    },
+    {
+      city: "kaukhali",
+      sub: "Kaukhali",
+      postalCode: "8510",
+    },
+    {
+      city: "kaukhali",
+      sub: "Keundia",
+      postalCode: "8511",
+    },
+    {
+      city: "kaukhali",
+      sub: "Joykul",
+      postalCode: "8512",
+    },
+    {
+      city: "kaukhali",
+      sub: "Jolagati",
+      postalCode: "8513",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Mathbaria",
+      postalCode: "8560",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Tushkhali",
+      postalCode: "8561",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Halta",
+      postalCode: "8562",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Gulishakhali",
+      postalCode: "8563",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Tiarkhali",
+      postalCode: "8564",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Betmor Natun Hat",
+      postalCode: "8565",
+    },
+    {
+      city: "Mathbaria",
+      sub: "Shilarganj",
+      postalCode: "8566",
+    },
+    {
+      city: "Nazirpur",
+      sub: "Nazirpur",
+      postalCode: "8540",
+    },
+    {
+      city: "Nazirpur",
+      sub: "Sriramkathi",
+      postalCode: "8541",
+    },
+    {
+      city: "Pirojpur Sadar",
+      sub: "Pirojpur Sadar",
+      postalCode: "8500",
+    },
+    {
+      city: "Pirojpur Sadar",
+      sub: "Hularhat",
+      postalCode: "8501",
+    },
+    {
+      city: "Pirojpur Sadar",
+      sub: "Parerhat",
+      postalCode: "8502",
+    },
+    {
+      city: "Swarupkathi",
+      sub: "Swarupkathi",
+      postalCode: "8520",
+    },
+    {
+      city: "Swarupkathi",
+      sub: "Darus Sunnat",
+      postalCode: "8521",
+    },
+    {
+      city: "Swarupkathi",
+      sub: "Kaurikhara",
+      postalCode: "8522",
+    },
+    {
+      city: "Swarupkathi",
+      sub: "Jalabari",
+      postalCode: "8523",
+    },
+  ],
+  Patuakhali: [
+    {
+      city: "Bauphal",
+      sub: "Bauphal",
+      postalCode: "8620",
+    },
+    {
+      city: "Bauphal",
+      sub: "Bagabandar",
+      postalCode: "8621",
+    },
+    {
+      city: "Bauphal",
+      sub: "Birpasha",
+      postalCode: "8622",
+    },
+    {
+      city: "Bauphal",
+      sub: "Kalishari",
+      postalCode: "8623",
+    },
+    {
+      city: "Bauphal",
+      sub: "Kalaia",
+      postalCode: "8624",
+    },
+    {
+      city: "Dashmina",
+      sub: "Dashmina",
+      postalCode: "8630",
+    },
+    {
+      city: "Galachipa",
+      sub: "Galachipa",
+      postalCode: "8640",
+    },
+    {
+      city: "Galachipa",
+      sub: "Gazipur Bandar",
+      postalCode: "8641",
+    },
+    {
+      city: "Khepupara",
+      sub: "Khepupara",
+      postalCode: "8650",
+    },
+    {
+      city: "Khepupara",
+      sub: "Mahipur",
+      postalCode: "8651",
+    },
+    {
+      city: "Patuakhali Sadar",
+      sub: "Patuakhali Sadar",
+      postalCode: "8600",
+    },
+    {
+      city: "Patuakhali Sadar",
+      sub: "Moukaran",
+      postalCode: "8601",
+    },
+    {
+      city: "Patuakhali Sadar",
+      sub: "Dumkee",
+      postalCode: "8602",
+    },
+    {
+      city: "Patuakhali Sadar",
+      sub: "Rahimabad",
+      postalCode: "8603",
+    },
+    {
+      city: "Subidkhali",
+      sub: "Subidkhali",
+      postalCode: "8610",
+    },
+  ],
+  Barguna: [
+    {
+      city: "Amtali",
+      sub: "Amtali",
+      postalCode: "8710",
+    },
+    {
+      city: "Bamna",
+      sub: "Bamna",
+      postalCode: "8730",
+    },
+    {
+      city: "Barguna Sadar",
+      sub: "Barguna Sadar",
+      postalCode: "8700",
+    },
+    {
+      city: "Barguna Sadar",
+      sub: "Nali Bandar",
+      postalCode: "8701",
+    },
+    {
+      city: "Betagi",
+      sub: "Betagi",
+      postalCode: "8740",
+    },
+    {
+      city: "Betagi",
+      sub: "Darul Ulam",
+      postalCode: "8741",
+    },
+    {
+      city: "Patharghata",
+      sub: "Patharghata",
+      postalCode: "8720",
+    },
+    {
+      city: "Patharghata",
+      sub: "Kakchira",
+      postalCode: "8721",
+    },
+  ],
+  Panchagarh: [
+    {
+      city: "Boda",
+      sub: "Boda",
+      postalCode: "5010",
+    },
+    {
+      city: "Chotto Dab",
+      sub: "Chotto Dab",
+      postalCode: "5040",
+    },
+    {
+      city: "Chotto Dab",
+      sub: "Mirjapur",
+      postalCode: "5041",
+    },
+    {
+      city: "Dabiganj",
+      sub: "Dabiganj",
+      postalCode: "5020",
+    },
+    {
+      city: "Panchagra Sadar",
+      sub: "Panchagar Sadar",
+      postalCode: "5000",
+    },
+    {
+      city: "Tetulia",
+      sub: "Tetulia",
+      postalCode: "5030",
+    },
+  ],
+  Thakurgaon: [
+    {
+      city: "Baliadangi",
+      sub: "Baliadangi",
+      postalCode: "5140",
+    },
+    {
+      city: "Baliadangi",
+      sub: "Lahiri",
+      postalCode: "5141",
+    },
+    {
+      city: "Jibanpur",
+      sub: "Jibanpur",
+      postalCode: "5130",
+    },
+    {
+      city: "Pirganj",
+      sub: "Pirganj",
+      postalCode: "5110",
+    },
+    {
+      city: "Pirganj",
+      sub: "Pirganj",
+      postalCode: "5470",
+    },
+    {
+      city: "Rani Sankail",
+      sub: "Rani Sankail",
+      postalCode: "5120",
+    },
+    {
+      city: "Rani Sankail",
+      sub: "Nekmarad",
+      postalCode: "5121",
+    },
+    {
+      city: "Thakurgaon Sadar",
+      sub: "Thakurgaon Sadar",
+      postalCode: "5100",
+    },
+    {
+      city: "Thakurgaon Sadar",
+      sub: "Thakurgaon Road",
+      postalCode: "5101",
+    },
+    {
+      city: "Thakurgaon Sadar",
+      sub: "Shibganj",
+      postalCode: "5102",
+    },
+    {
+      city: "Thakurgaon Sadar",
+      sub: "Ruhia",
+      postalCode: "5103",
+    },
+  ],
+  Dinajpur: [
+    {
+      city: "Bangla Hili",
+      sub: "Bangla Hili",
+      postalCode: "5270",
+    },
+    {
+      city: "Biral",
+      sub: "Biral",
+      postalCode: "5210",
+    },
+    {
+      city: "Birampur",
+      sub: "Birampur",
+      postalCode: "5266",
+    },
+    {
+      city: "Birganj",
+      sub: "Birganj",
+      postalCode: "5220",
+    },
+    {
+      city: "Chrirbandar",
+      sub: "Chrirbandar",
+      postalCode: "5240",
+    },
+    {
+      city: "Chrirbandar",
+      sub: "Ranirbandar",
+      postalCode: "5241",
+    },
+    {
+      city: "Dinajpur Sadar",
+      sub: "Dinajpur Sadar",
+      postalCode: "5200",
+    },
+    {
+      city: "Dinajpur Sadar",
+      sub: "Dinajpur Rajbari",
+      postalCode: "5201",
+    },
+    {
+      city: "Khansama",
+      sub: "Khansama",
+      postalCode: "5230",
+    },
+    {
+      city: "Khansama",
+      sub: "Pakarhat",
+      postalCode: "5231",
+    },
+    {
+      city: "Maharajganj",
+      sub: "Maharajganj",
+      postalCode: "5226",
+    },
+    {
+      city: "Nababganj",
+      sub: "Nababganj",
+      postalCode: "5280",
+    },
+    {
+      city: "Nababganj",
+      sub: "Daudpur",
+      postalCode: "5281",
+    },
+    {
+      city: "Nababganj",
+      sub: "Gopalpur",
+      postalCode: "5282",
+    },
+    {
+      city: "Osmanpur",
+      sub: "Osmanpur",
+      postalCode: "5290",
+    },
+    {
+      city: "Osmanpur",
+      sub: "Ghoraghat",
+      postalCode: "5291",
+    },
+    {
+      city: "Parbatipur",
+      sub: "Parbatipur",
+      postalCode: "5250",
+    },
+    {
+      city: "Phulbari",
+      sub: "Phulbari",
+      postalCode: "5260",
+    },
+    {
+      city: "Setabganj",
+      sub: "Setabganj",
+      postalCode: "5216",
+    },
+  ],
+  Nilphamari: [
+    {
+      city: "Dimla",
+      sub: "Dimla",
+      postalCode: "5350",
+    },
+    {
+      city: "Dimla",
+      sub: "Ghaga Kharibari",
+      postalCode: "5351",
+    },
+    {
+      city: "Domar",
+      sub: "Domar",
+      postalCode: "5340",
+    },
+    {
+      city: "Domar",
+      sub: "Chilahati",
+      postalCode: "5341",
+    },
+    {
+      city: "Jaldhaka",
+      sub: "Jaldhaka",
+      postalCode: "5330",
+    },
+    {
+      city: "Kishoriganj",
+      sub: "Kishoriganj",
+      postalCode: "5320",
+    },
+    {
+      city: "Nilphamari Sadar",
+      sub: "Nilphamari Sadar",
+      postalCode: "5300",
+    },
+    {
+      city: "Nilphamari Sadar",
+      sub: "Nilphamari Sugar Mil",
+      postalCode: "5301",
+    },
+    {
+      city: "Syedpur",
+      sub: "Syedpur",
+      postalCode: "5310",
+    },
+    {
+      city: "Syedpur",
+      sub: "Syedpur Upashahar",
+      postalCode: "5311",
+    },
+  ],
+  Rangpur: [
+    {
+      city: "Badarganj",
+      sub: "Badarganj",
+      postalCode: "5430",
+    },
+    {
+      city: "Badarganj",
+      sub: "Shyampur",
+      postalCode: "5431",
+    },
+    {
+      city: "Gangachara",
+      sub: "Gangachara",
+      postalCode: "5410",
+    },
+    {
+      city: "Kaunia",
+      sub: "Kaunia",
+      postalCode: "5440",
+    },
+    {
+      city: "Kaunia",
+      sub: "Haragachh",
+      postalCode: "5441",
+    },
+    {
+      city: "Mithapukur",
+      sub: "Mithapukur",
+      postalCode: "5460",
+    },
+    {
+      city: "Pirgachha",
+      sub: "Pirgachha",
+      postalCode: "5450",
+    },
+    {
+      city: "Rangpur Sadar",
+      sub: "Rangpur Sadar",
+      postalCode: "5400",
+    },
+    {
+      city: "Rangpur Sadar",
+      sub: "Rangpur Upa-Shahar",
+      postalCode: "5401",
+    },
+    {
+      city: "Rangpur Sadar",
+      sub: "Alamnagar",
+      postalCode: "5402",
+    },
+    {
+      city: "Rangpur Sadar",
+      sub: "Mahiganj",
+      postalCode: "5403",
+    },
+    {
+      city: "Rangpur Sadar",
+      sub: "Rangpur Cadet Colleg",
+      postalCode: "5404",
+    },
+    {
+      city: "Rangpur Sadar",
+      sub: "Rangpur Carmiecal Col",
+      postalCode: "5405",
+    },
+    {
+      city: "Taraganj",
+      sub: "Taraganj",
+      postalCode: "5420",
+    },
+  ],
+  Lalmonirhat: [
+    {
+      city: "Aditmari",
+      sub: "Aditmari",
+      postalCode: "5510",
+    },
+    {
+      city: "Hatibandha",
+      sub: "Hatibandha",
+      postalCode: "5530",
+    },
+    {
+      city: "Lalmonirhat Sadar",
+      sub: "Lalmonirhat Sadar",
+      postalCode: "5500",
+    },
+    {
+      city: "Lalmonirhat Sadar",
+      sub: "Moghalhat",
+      postalCode: "5501",
+    },
+    {
+      city: "Lalmonirhat Sadar",
+      sub: "Kulaghat SO",
+      postalCode: "5502",
+    },
+    {
+      city: "Patgram",
+      sub: "Patgram",
+      postalCode: "5540",
+    },
+    {
+      city: "Patgram",
+      sub: "Baura",
+      postalCode: "5541",
+    },
+    {
+      city: "Patgram",
+      sub: "Burimari",
+      postalCode: "5542",
+    },
+    {
+      city: "Tushbhandar",
+      sub: "Tushbhandar",
+      postalCode: "5520",
+    },
+  ],
+  Kurigram: [
+    {
+      city: "Bhurungamari",
+      sub: "Bhurungamari",
+      postalCode: "5670",
+    },
+    {
+      city: "Chilmari",
+      sub: "Chilmari",
+      postalCode: "5630",
+    },
+    {
+      city: "Chilmari",
+      sub: "Jorgachh",
+      postalCode: "5631",
+    },
+    {
+      city: "Kurigram Sadar",
+      sub: "Kurigram Sadar",
+      postalCode: "5600",
+    },
+    {
+      city: "Kurigram Sadar",
+      sub: "Pandul",
+      postalCode: "5601",
+    },
+    {
+      city: "Kurigram Sadar",
+      sub: "Phulbari",
+      postalCode: "5680",
+    },
+    {
+      city: "Nageshwar",
+      sub: "Nageshwar",
+      postalCode: "5660",
+    },
+    {
+      city: "Rajarhat",
+      sub: "Rajarhat",
+      postalCode: "5610",
+    },
+    {
+      city: "Rajarhat",
+      sub: "Nazimkhan",
+      postalCode: "5611",
+    },
+    {
+      city: "Rajibpur",
+      sub: "Rajibpur",
+      postalCode: "5650",
+    },
+    {
+      city: "Roumari",
+      sub: "Roumari",
+      postalCode: "5640",
+    },
+    {
+      city: "Ulipur",
+      sub: "Ulipur",
+      postalCode: "5620",
+    },
+    {
+      city: "Ulipur",
+      sub: "Bazarhat",
+      postalCode: "5621",
+    },
+  ],
+  Gaibandha: [
+    {
+      city: "Bonarpara",
+      sub: "Bonarpara",
+      postalCode: "5750",
+    },
+    {
+      city: "Bonarpara",
+      sub: "saghata",
+      postalCode: "5751",
+    },
+    {
+      city: "Gaibandha Sadar",
+      sub: "Gaibandha Sadar",
+      postalCode: "5700",
+    },
+    {
+      city: "Gobindaganj",
+      sub: "Gobindhaganj",
+      postalCode: "5740",
+    },
+    {
+      city: "Gobindaganj",
+      sub: "Mahimaganj",
+      postalCode: "5741",
+    },
+    {
+      city: "Palashbari",
+      sub: "Palashbari",
+      postalCode: "5730",
+    },
+    {
+      city: "Phulchhari",
+      sub: "Phulchhari",
+      postalCode: "5760",
+    },
+    {
+      city: "Phulchhari",
+      sub: "Bharatkhali",
+      postalCode: "5761",
+    },
+    {
+      city: "Saadullapur",
+      sub: "Saadullapur",
+      postalCode: "5710",
+    },
+    {
+      city: "Saadullapur",
+      sub: "Naldanga",
+      postalCode: "5711",
+    },
+    {
+      city: "Sundarganj",
+      sub: "Sundarganj",
+      postalCode: "5720",
+    },
+    {
+      city: "Sundarganj",
+      sub: "Bamandanga",
+      postalCode: "5721",
+    },
+  ],
+  Bogra: [
+    {
+      city: "Alamdighi",
+      sub: "Adamdighi",
+      postalCode: "5890",
+    },
+    {
+      city: "Alamdighi",
+      sub: "Santahar",
+      postalCode: "5891",
+    },
+    {
+      city: "Alamdighi",
+      sub: "Nasharatpur",
+      postalCode: "5892",
+    },
+    {
+      city: "Bogra Sadar",
+      sub: "Bogra Sadar",
+      postalCode: "5800",
+    },
+    {
+      city: "Bogra Sadar",
+      sub: "Bogra Canttonment",
+      postalCode: "5801",
+    },
+    {
+      city: "Dhunat",
+      sub: "Dhunat",
+      postalCode: "5850",
+    },
+    {
+      city: "Dhunat",
+      sub: "Gosaibari",
+      postalCode: "5851",
+    },
+    {
+      city: "Dupchachia",
+      sub: "Dupchachia",
+      postalCode: "5880",
+    },
+    {
+      city: "Dupchachia",
+      sub: "Talora",
+      postalCode: "5881",
+    },
+    {
+      city: "Gabtoli",
+      sub: "Gabtoli",
+      postalCode: "5820",
+    },
+    {
+      city: "Gabtoli",
+      sub: "Sukhanpukur",
+      postalCode: "5821",
+    },
+    {
+      city: "Kahalu",
+      sub: "Kahalu",
+      postalCode: "5870",
+    },
+    {
+      city: "Nandigram",
+      sub: "Nandigram",
+      postalCode: "5860",
+    },
+    {
+      city: "Sariakandi",
+      sub: "Sariakandi",
+      postalCode: "5830",
+    },
+    {
+      city: "Sariakandi",
+      sub: "Chandan Baisha",
+      postalCode: "5831",
+    },
+    {
+      city: "Sherpur",
+      sub: "Sherpur",
+      postalCode: "5840",
+    },
+    {
+      city: "Sherpur",
+      sub: "Chandaikona",
+      postalCode: "5841",
+    },
+    {
+      city: "Sherpur",
+      sub: "Palli Unnyan Accadem",
+      postalCode: "5842",
+    },
+    {
+      city: "Shibganj",
+      sub: "Shibganj",
+      postalCode: "5810",
+    },
+    {
+      city: "Sonatola",
+      sub: "Sonatola",
+      postalCode: "5826",
+    },
+  ],
+  Joypurhat: [
+    {
+      city: "Akkelpur",
+      sub: "Akklepur",
+      postalCode: "5940",
+    },
+    {
+      city: "Akkelpur",
+      sub: "jamalganj",
+      postalCode: "5941",
+    },
+    {
+      city: "Akkelpur",
+      sub: "Tilakpur",
+      postalCode: "5942",
+    },
+    {
+      city: "Joypurhat Sadar",
+      sub: "Joypurhat Sadar",
+      postalCode: "5900",
+    },
+    {
+      city: "kalai",
+      sub: "kalai",
+      postalCode: "5930",
+    },
+    {
+      city: "Khetlal",
+      sub: "Khetlal",
+      postalCode: "5920",
+    },
+    {
+      city: "panchbibi",
+      sub: "Panchbibi",
+      postalCode: "5910",
+    },
+  ],
+  Rajshahi: [
+    {
+      city: "Bagha",
+      sub: "Bagha",
+      postalCode: "6280",
+    },
+    {
+      city: "Bagha",
+      sub: "Arani",
+      postalCode: "6281",
+    },
+    {
+      city: "Bhabaniganj",
+      sub: "Bhabaniganj",
+      postalCode: "6250",
+    },
+    {
+      city: "Bhabaniganj",
+      sub: "Taharpur",
+      postalCode: "6251",
+    },
+    {
+      city: "Charghat",
+      sub: "Charghat",
+      postalCode: "6270",
+    },
+    {
+      city: "Charghat",
+      sub: "Sarda",
+      postalCode: "6271",
+    },
+    {
+      city: "Durgapur",
+      sub: "Durgapur",
+      postalCode: "6240",
+    },
+    {
+      city: "Godagari",
+      sub: "Godagari",
+      postalCode: "6290",
+    },
+    {
+      city: "Godagari",
+      sub: "Premtoli",
+      postalCode: "6291",
+    },
+    {
+      city: "Khod Mohanpur",
+      sub: "Khodmohanpur",
+      postalCode: "6220",
+    },
+    {
+      city: "Lalitganj",
+      sub: "Lalitganj",
+      postalCode: "6210",
+    },
+    {
+      city: "Lalitganj",
+      sub: "Rajshahi Sugar Mills",
+      postalCode: "6211",
+    },
+    {
+      city: "Lalitganj",
+      sub: "Shyampur",
+      postalCode: "6212",
+    },
+    {
+      city: "Putia",
+      sub: "Putia",
+      postalCode: "6260",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Rajshahi Sadar",
+      postalCode: "6000",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Ghuramara",
+      postalCode: "6100",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Rajshahi Court",
+      postalCode: "6201",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Rajshahi Canttonment",
+      postalCode: "6202",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Sapura",
+      postalCode: "6203",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Kazla",
+      postalCode: "6204",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Rajshahi University",
+      postalCode: "6205",
+    },
+    {
+      city: "Rajshahi Sadar",
+      sub: "Binodpur Bazar",
+      postalCode: "6206",
+    },
+    {
+      city: "Tanor",
+      sub: "Tanor",
+      postalCode: "6230",
+    },
+  ],
+  Chapinawabganj: [
+    {
+      city: "Bholahat",
+      sub: "Bholahat",
+      postalCode: "6330",
+    },
+    {
+      city: "Chapinawabganj Sadar",
+      sub: "Chapinawbganj Sadar",
+      postalCode: "6300",
+    },
+    {
+      city: "Chapinawabganj Sadar",
+      sub: "Rajarampur",
+      postalCode: "6301",
+    },
+    {
+      city: "Chapinawabganj Sadar",
+      sub: "Ramchandrapur",
+      postalCode: "6302",
+    },
+    {
+      city: "Chapinawabganj Sadar",
+      sub: "Amnura",
+      postalCode: "6303",
+    },
+    {
+      city: "Nachol",
+      sub: "Nachol",
+      postalCode: "6310",
+    },
+    {
+      city: "Nachol",
+      sub: "Mandumala",
+      postalCode: "6311",
+    },
+    {
+      city: "Rohanpur",
+      sub: "Rohanpur",
+      postalCode: "6320",
+    },
+    {
+      city: "Rohanpur",
+      sub: "Gomashtapur",
+      postalCode: "6321",
+    },
+    {
+      city: "Shibganj U.P.O",
+      sub: "Shibganj U.P.O",
+      postalCode: "6340",
+    },
+    {
+      city: "Shibganj U.P.O",
+      sub: "Kansart",
+      postalCode: "6341",
+    },
+    {
+      city: "Shibganj U.P.O",
+      sub: "Manaksha",
+      postalCode: "6342",
+    },
+  ],
+  Natore: [
+    {
+      city: "Gopalpur UPO",
+      sub: "Gopalpur U.P.O",
+      postalCode: "6420",
+    },
+    {
+      city: "Gopalpur UPO",
+      sub: "Lalpur S.O",
+      postalCode: "6421",
+    },
+    {
+      city: "Gopalpur UPO",
+      sub: "Abdulpur",
+      postalCode: "6422",
+    },
+    {
+      city: "Harua",
+      sub: "Harua",
+      postalCode: "6430",
+    },
+    {
+      city: "Harua",
+      sub: "Dayarampur",
+      postalCode: "6431",
+    },
+    {
+      city: "Harua",
+      sub: "Baraigram",
+      postalCode: "6432",
+    },
+    {
+      city: "Hatgurudaspur",
+      sub: "Hatgurudaspur",
+      postalCode: "6440",
+    },
+    {
+      city: "Laxman",
+      sub: "Laxman",
+      postalCode: "6410",
+    },
+    {
+      city: "Natore Sadar",
+      sub: "Natore Sadar",
+      postalCode: "6400",
+    },
+    {
+      city: "Natore Sadar",
+      sub: "Digapatia",
+      postalCode: "6401",
+    },
+    {
+      city: "Natore Sadar",
+      sub: "Baiddyabal Gharia",
+      postalCode: "6402",
+    },
+    {
+      city: "Natore Sadar",
+      sub: "Madhnagar",
+      postalCode: "6403",
+    },
+    {
+      city: "Singra",
+      sub: "Singra",
+      postalCode: "6450",
+    },
+  ],
+  Naogaon: [
+    {
+      city: "Ahsanganj",
+      sub: "Ahsanganj",
+      postalCode: "6596",
+    },
+    {
+      city: "Ahsanganj",
+      sub: "Bandai",
+      postalCode: "6597",
+    },
+    {
+      city: "Badalgachhi",
+      sub: "Badalgachhi",
+      postalCode: "6570",
+    },
+    {
+      city: "Dhamuirhat",
+      sub: "Dhamuirhat",
+      postalCode: "6580",
+    },
+    {
+      city: "Mahadebpur",
+      sub: "Mahadebpur",
+      postalCode: "6530",
+    },
+    {
+      city: "Naogaon Sadar",
+      sub: "Naogaon Sadar",
+      postalCode: "6500",
+    },
+    {
+      city: "Niamatpur",
+      sub: "Niamatpur",
+      postalCode: "6520",
+    },
+    {
+      city: "Nitpur",
+      sub: "Nitpur",
+      postalCode: "6550",
+    },
+    {
+      city: "Nitpur",
+      sub: "Porsa",
+      postalCode: "6551",
+    },
+    {
+      city: "Nitpur",
+      sub: "Panguria",
+      postalCode: "6552",
+    },
+    {
+      city: "Patnitala",
+      sub: "Patnitala",
+      postalCode: "6540",
+    },
+    {
+      city: "Prasadpur",
+      sub: "Prasadpur",
+      postalCode: "6510",
+    },
+    {
+      city: "Prasadpur",
+      sub: "Manda",
+      postalCode: "6511",
+    },
+    {
+      city: "Prasadpur",
+      sub: "Balihar",
+      postalCode: "6512",
+    },
+    {
+      city: "Raninagar",
+      sub: "Raninagar",
+      postalCode: "6590",
+    },
+    {
+      city: "Raninagar",
+      sub: "Kashimpur",
+      postalCode: "6591",
+    },
+    {
+      city: "Sapahar",
+      sub: "Sapahar",
+      postalCode: "6560",
+    },
+    {
+      city: "Sapahar",
+      sub: "Moduhil",
+      postalCode: "6561",
+    },
+  ],
+  Pabna: [
+    {
+      city: "Banwarinagar",
+      sub: "Banwarinagar",
+      postalCode: "6650",
+    },
+    {
+      city: "Bera",
+      sub: "Bera",
+      postalCode: "6680",
+    },
+    {
+      city: "Bera",
+      sub: "Nakalia",
+      postalCode: "6681",
+    },
+    {
+      city: "Bera",
+      sub: "Kashinathpur",
+      postalCode: "6682",
+    },
+    {
+      city: "Bera",
+      sub: "Puran Bharenga",
+      postalCode: "6683",
+    },
+    {
+      city: "Bhangura",
+      sub: "Bhangura",
+      postalCode: "6640",
+    },
+    {
+      city: "Chatmohar",
+      sub: "Chatmohar",
+      postalCode: "6630",
+    },
+    {
+      city: "Debottar",
+      sub: "Debottar",
+      postalCode: "6610",
+    },
+    {
+      city: "Ishwardi",
+      sub: "Ishwardi",
+      postalCode: "6620",
+    },
+    {
+      city: "Ishwardi",
+      sub: "Dhapari",
+      postalCode: "6621",
+    },
+    {
+      city: "Ishwardi",
+      sub: "Pakshi",
+      postalCode: "6622",
+    },
+    {
+      city: "Ishwardi",
+      sub: "Rajapur",
+      postalCode: "6623",
+    },
+    {
+      city: "Pabna Sadar",
+      sub: "Pabna Sadar",
+      postalCode: "6600",
+    },
+    {
+      city: "Pabna Sadar",
+      sub: "Kaliko Cotton Mills",
+      postalCode: "6601",
+    },
+    {
+      city: "Pabna Sadar",
+      sub: "Hamayetpur",
+      postalCode: "6602",
+    },
+    {
+      city: "Sathia",
+      sub: "Sathia",
+      postalCode: "6670",
+    },
+    {
+      city: "Sujanagar",
+      sub: "Sujanagar",
+      postalCode: "6660",
+    },
+    {
+      city: "Sujanagar",
+      sub: "Sagarkandi",
+      postalCode: "6661",
+    },
+  ],
+  Sirajganj: [
+    {
+      city: "Baiddya Jam Toil",
+      sub: "Baiddya Jam Toil",
+      postalCode: "6730",
+    },
+    {
+      city: "Belkuchi",
+      sub: "Belkuchi",
+      postalCode: "6740",
+    },
+    {
+      city: "Belkuchi",
+      sub: "Sohagpur",
+      postalCode: "6741",
+    },
+    {
+      city: "Belkuchi",
+      sub: "Rajapur",
+      postalCode: "6742",
+    },
+    {
+      city: "Belkuchi",
+      sub: "Enayetpur",
+      postalCode: "6751",
+    },
+    {
+      city: "Belkuchi",
+      sub: "Sthal",
+      postalCode: "6752",
+    },
+    {
+      city: "Dhangora",
+      sub: "Dhangora",
+      postalCode: "6720",
+    },
+    {
+      city: "Dhangora",
+      sub: "Malonga",
+      postalCode: "6721",
+    },
+    {
+      city: "Kazipur",
+      sub: "Kazipur",
+      postalCode: "6710",
+    },
+    {
+      city: "Kazipur",
+      sub: "Shuvgachha",
+      postalCode: "6711",
+    },
+    {
+      city: "Kazipur",
+      sub: "Gandail",
+      postalCode: "6712",
+    },
+    {
+      city: "Shahjadpur",
+      sub: "Shahjadpur",
+      postalCode: "6770",
+    },
+    {
+      city: "Shahjadpur",
+      sub: "Porjana",
+      postalCode: "6771",
+    },
+    {
+      city: "Shahjadpur",
+      sub: "Jamirta",
+      postalCode: "6772",
+    },
+    {
+      city: "Shahjadpur",
+      sub: "Kaijuri",
+      postalCode: "6773",
+    },
+    {
+      city: "Sirajganj Sadar",
+      sub: "Sirajganj Sadar",
+      postalCode: "6700",
+    },
+    {
+      city: "Sirajganj Sadar",
+      sub: "Raipur",
+      postalCode: "6701",
+    },
+    {
+      city: "Sirajganj Sadar",
+      sub: "Rashidabad",
+      postalCode: "6702",
+    },
+    {
+      city: "Tarash",
+      sub: "Tarash",
+      postalCode: "6780",
+    },
+    {
+      city: "Ullapara",
+      sub: "Ullapara",
+      postalCode: "6760",
+    },
+    {
+      city: "Ullapara",
+      sub: "Ullapara R.S",
+      postalCode: "6761",
+    },
+    {
+      city: "Ullapara",
+      sub: "Lahiri Mohanpur",
+      postalCode: "6762",
+    },
+    {
+      city: "Ullapara",
+      sub: "Salap",
+      postalCode: "6763",
+    },
+  ],
+  Sunamganj: [
+    {
+      city: "Bishamsarpur",
+      sub: "Bishamsapur",
+      postalCode: "3010",
+    },
+    {
+      city: "Chhatak",
+      sub: "Chhatak",
+      postalCode: "3080",
+    },
+    {
+      city: "Chhatak",
+      sub: "Chhatak Cement Facto",
+      postalCode: "3081",
+    },
+    {
+      city: "Chhatak",
+      sub: "Chhatak Paper Mills",
+      postalCode: "3082",
+    },
+    {
+      city: "Chhatak",
+      sub: "Gabindaganj",
+      postalCode: "3083",
+    },
+    {
+      city: "Chhatak",
+      sub: "Gabindaganj Natun Ba",
+      postalCode: "3084",
+    },
+    {
+      city: "Chhatak",
+      sub: "Khurma",
+      postalCode: "3085",
+    },
+    {
+      city: "Chhatak",
+      sub: "Moinpur",
+      postalCode: "3086",
+    },
+    {
+      city: "Chhatak",
+      sub: "jahidpur",
+      postalCode: "3087",
+    },
+    {
+      city: "Chhatak",
+      sub: "Islamabad",
+      postalCode: "3088",
+    },
+    {
+      city: "Chhatak",
+      sub: "Chourangi Bazar",
+      postalCode: "3893",
+    },
+    {
+      city: "Dhirai Chandpur",
+      sub: "Dhirai Chandpur",
+      postalCode: "3040",
+    },
+    {
+      city: "Dhirai Chandpur",
+      sub: "Jagdal",
+      postalCode: "3041",
+    },
+    {
+      city: "Duara bazar",
+      sub: "Duara bazar",
+      postalCode: "3070",
+    },
+    {
+      city: "Ghungiar",
+      sub: "Ghungiar",
+      postalCode: "3050",
+    },
+    {
+      city: "Jagnnathpur",
+      sub: "Jagnnathpur",
+      postalCode: "3060",
+    },
+    {
+      city: "Jagnnathpur",
+      sub: "Syedpur",
+      postalCode: "3061",
+    },
+    {
+      city: "Jagnnathpur",
+      sub: "Atuajan",
+      postalCode: "3062",
+    },
+    {
+      city: "Jagnnathpur",
+      sub: "Hasan Fatemapur",
+      postalCode: "3063",
+    },
+    {
+      city: "Jagnnathpur",
+      sub: "Rasulganj",
+      postalCode: "3064",
+    },
+    {
+      city: "Jagnnathpur",
+      sub: "Shiramsi",
+      postalCode: "3065",
+    },
+    {
+      city: "Sachna",
+      sub: "Sachna",
+      postalCode: "3020",
+    },
+    {
+      city: "Sunamganj Sadar",
+      sub: "Sunamganj Sadar",
+      postalCode: "3000",
+    },
+    {
+      city: "Sunamganj Sadar",
+      sub: "Pagla",
+      postalCode: "3001",
+    },
+    {
+      city: "Sunamganj Sadar",
+      sub: "Patharia",
+      postalCode: "3002",
+    },
+    {
+      city: "Tahirpur",
+      sub: "Tahirpur",
+      postalCode: "3030",
+    },
+  ],
+  Sylhet: [
+    {
+      city: "Balaganj",
+      sub: "Balaganj",
+      postalCode: "3120",
+    },
+    {
+      city: "Balaganj",
+      sub: "Karua",
+      postalCode: "3121",
+    },
+    {
+      city: "Balaganj",
+      sub: "Brahman Shashon",
+      postalCode: "3122",
+    },
+    {
+      city: "Balaganj",
+      sub: "Tajpur",
+      postalCode: "3123",
+    },
+    {
+      city: "Balaganj",
+      sub: "Goala Bazar",
+      postalCode: "3124",
+    },
+    {
+      city: "Balaganj",
+      sub: "Begumpur",
+      postalCode: "3125",
+    },
+    {
+      city: "Balaganj",
+      sub: "Omarpur",
+      postalCode: "3126",
+    },
+    {
+      city: "Balaganj",
+      sub: "Kathal Khair",
+      postalCode: "3127",
+    },
+    {
+      city: "Balaganj",
+      sub: "Gaharpur",
+      postalCode: "3128",
+    },
+    {
+      city: "Balaganj",
+      sub: "Natun Bazar",
+      postalCode: "3129",
+    },
+    {
+      city: "Bianibazar",
+      sub: "Bianibazar",
+      postalCode: "3170",
+    },
+    {
+      city: "Bianibazar",
+      sub: "jaldup",
+      postalCode: "3171",
+    },
+    {
+      city: "Bianibazar",
+      sub: "Mathiura",
+      postalCode: "3172",
+    },
+    {
+      city: "Bianibazar",
+      sub: "Kurar bazar",
+      postalCode: "3173",
+    },
+    {
+      city: "Bianibazar",
+      sub: "Salia bazar",
+      postalCode: "3174",
+    },
+    {
+      city: "Bianibazar",
+      sub: "Churkai",
+      postalCode: "3175",
+    },
+    {
+      city: "Bishwanath",
+      sub: "Bishwanath",
+      postalCode: "3130",
+    },
+    {
+      city: "Bishwanath",
+      sub: "Dashghar",
+      postalCode: "3131",
+    },
+    {
+      city: "Bishwanath",
+      sub: "Doulathpur",
+      postalCode: "3132",
+    },
+    {
+      city: "Bishwanath",
+      sub: "Deokalas",
+      postalCode: "3133",
+    },
+    {
+      city: "Bishwanath",
+      sub: "Singer kanch",
+      postalCode: "3134",
+    },
+    {
+      city: "Fenchuganj",
+      sub: "Fenchuganj",
+      postalCode: "3116",
+    },
+    {
+      city: "Fenchuganj",
+      sub: "Fenchuganj SareKarkh",
+      postalCode: "3117",
+    },
+    {
+      city: "Goainhat",
+      sub: "Goainhat",
+      postalCode: "3150",
+    },
+    {
+      city: "Goainhat",
+      sub: "Jaflong",
+      postalCode: "3151",
+    },
+    {
+      city: "Goainhat",
+      sub: "Chiknagul",
+      postalCode: "3152",
+    },
+    {
+      city: "Gopalganj",
+      sub: "Gopalgannj",
+      postalCode: "3160",
+    },
+    {
+      city: "Gopalganj",
+      sub: "Dhaka Dakkhin",
+      postalCode: "3161",
+    },
+    {
+      city: "Gopalganj",
+      sub: "Dakkhin Bhadashore",
+      postalCode: "3162",
+    },
+    {
+      city: "Gopalganj",
+      sub: "Ranaping",
+      postalCode: "3163",
+    },
+    {
+      city: "Gopalganj",
+      sub: "banigram",
+      postalCode: "3164",
+    },
+    {
+      city: "Gopalganj",
+      sub: "Chandanpur",
+      postalCode: "3165",
+    },
+    {
+      city: "Jaintapur",
+      sub: "Jainthapur",
+      postalCode: "3156",
+    },
+    {
+      city: "Jakiganj",
+      sub: "Jakiganj",
+      postalCode: "3190",
+    },
+    {
+      city: "Jakiganj",
+      sub: "Ichhamati",
+      postalCode: "3191",
+    },
+    {
+      city: "Kanaighat",
+      sub: "Kanaighat",
+      postalCode: "3180",
+    },
+    {
+      city: "Kanaighat",
+      sub: "Chatulbazar",
+      postalCode: "3181",
+    },
+    {
+      city: "Kanaighat",
+      sub: "Manikganj",
+      postalCode: "3182",
+    },
+    {
+      city: "Kanaighat",
+      sub: "Gachbari",
+      postalCode: "3183",
+    },
+    {
+      city: "Kompanyganj",
+      sub: "Kompanyganj",
+      postalCode: "3140",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Sylhe Sadar",
+      postalCode: "3100",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Sylhet Cadet Col",
+      postalCode: "3101",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Sylhet Biman Bondar",
+      postalCode: "3102",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Khadimnagar",
+      postalCode: "3103",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Jalalabad Cantoment",
+      postalCode: "3104",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Silam",
+      postalCode: "3105",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Birahimpur",
+      postalCode: "3106",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Jalalabad",
+      postalCode: "3107",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Mogla",
+      postalCode: "3108",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Ranga Hajiganj",
+      postalCode: "3109",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Kadamtali",
+      postalCode: "3111",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Kamalbazer",
+      postalCode: "3112",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Lalbazar",
+      postalCode: "3113",
+    },
+    {
+      city: "Sylhet Sadar",
+      sub: "Shahajalal Science &",
+      postalCode: "3114",
+    },
+  ],
+  Moulvibazar: [
+    {
+      city: "Baralekha",
+      sub: "Baralekha",
+      postalCode: "3250",
+    },
+    {
+      city: "Baralekha",
+      sub: "Juri",
+      postalCode: "3251",
+    },
+    {
+      city: "Baralekha",
+      sub: "Dhakkhinbag",
+      postalCode: "3252",
+    },
+    {
+      city: "Baralekha",
+      sub: "Purbashahabajpur",
+      postalCode: "3253",
+    },
+    {
+      city: "Kamalganj",
+      sub: "Kamalganj",
+      postalCode: "3220",
+    },
+    {
+      city: "Kamalganj",
+      sub: "Keramatnaga",
+      postalCode: "3221",
+    },
+    {
+      city: "Kamalganj",
+      sub: "Patrakhola",
+      postalCode: "3222",
+    },
+    {
+      city: "Kamalganj",
+      sub: "Shamsher Nagar",
+      postalCode: "3223",
+    },
+    {
+      city: "Kamalganj",
+      sub: "Munshibazar",
+      postalCode: "3224",
+    },
+    {
+      city: "Kulaura",
+      sub: "Kulaura",
+      postalCode: "3230",
+    },
+    {
+      city: "Kulaura",
+      sub: "Tillagaon",
+      postalCode: "3231",
+    },
+    {
+      city: "Kulaura",
+      sub: "Langla",
+      postalCode: "3232",
+    },
+    {
+      city: "Kulaura",
+      sub: "Prithimpasha",
+      postalCode: "3233",
+    },
+    {
+      city: "Kulaura",
+      sub: "Kajaldhara",
+      postalCode: "3234",
+    },
+    {
+      city: "Kulaura",
+      sub: "Karimpur",
+      postalCode: "3235",
+    },
+    {
+      city: "Kulaura",
+      sub: "Baramchal",
+      postalCode: "3237",
+    },
+    {
+      city: "Moulvibazar Sadar",
+      sub: "Moulvibazar Sadar",
+      postalCode: "3200",
+    },
+    {
+      city: "Moulvibazar Sadar",
+      sub: "Barakapan",
+      postalCode: "3201",
+    },
+    {
+      city: "Moulvibazar Sadar",
+      sub: "Monumukh",
+      postalCode: "3202",
+    },
+    {
+      city: "Moulvibazar Sadar",
+      sub: "Afrozganj",
+      postalCode: "3203",
+    },
+    {
+      city: "Rajnagar",
+      sub: "Rajnagar",
+      postalCode: "3240",
+    },
+    {
+      city: "Srimangal",
+      sub: "Srimangal",
+      postalCode: "3210",
+    },
+    {
+      city: "Srimangal",
+      sub: "Narain Chora",
+      postalCode: "3211",
+    },
+    {
+      city: "Srimangal",
+      sub: "Kalighat",
+      postalCode: "3212",
+    },
+    {
+      city: "Srimangal",
+      sub: "Khejurichhara",
+      postalCode: "3213",
+    },
+    {
+      city: "Srimangal",
+      sub: "Satgaon",
+      postalCode: "3214",
+    },
+  ],
+  Hobiganj: [
+    {
+      city: "Azmireeganj",
+      sub: "Azmireeganj",
+      postalCode: "3360",
+    },
+    {
+      city: "Bahubal",
+      sub: "Bahubal",
+      postalCode: "3310",
+    },
+    {
+      city: "Baniachang",
+      sub: "Baniachang",
+      postalCode: "3350",
+    },
+    {
+      city: "Baniachang",
+      sub: "Jatrapasha",
+      postalCode: "3351",
+    },
+    {
+      city: "Baniachang",
+      sub: "Kadirganj",
+      postalCode: "3352",
+    },
+    {
+      city: "Chunarughat",
+      sub: "Chunarughat",
+      postalCode: "3320",
+    },
+    {
+      city: "Chunarughat",
+      sub: "Chandpurbagan",
+      postalCode: "3321",
+    },
+    {
+      city: "Chunarughat",
+      sub: "Narapati",
+      postalCode: "3322",
+    },
+    {
+      city: "Hobiganj Sadar",
+      sub: "Hobiganj Sadar",
+      postalCode: "3300",
+    },
+    {
+      city: "Hobiganj Sadar",
+      sub: "Shaestaganj",
+      postalCode: "3301",
+    },
+    {
+      city: "Hobiganj Sadar",
+      sub: "Gopaya",
+      postalCode: "3302",
+    },
+    {
+      city: "Kalauk",
+      sub: "Kalauk",
+      postalCode: "3340",
+    },
+    {
+      city: "Kalauk",
+      sub: "Lakhai",
+      postalCode: "3341",
+    },
+    {
+      city: "Madhabpur",
+      sub: "Madhabpur",
+      postalCode: "3330",
+    },
+    {
+      city: "Madhabpur",
+      sub: "Itakhola",
+      postalCode: "3331",
+    },
+    {
+      city: "Madhabpur",
+      sub: "Shahajibazar",
+      postalCode: "3332",
+    },
+    {
+      city: "Madhabpur",
+      sub: "Saihamnagar",
+      postalCode: "3333",
+    },
+    {
+      city: "Nabiganj",
+      sub: "Nabiganj",
+      postalCode: "3370",
+    },
+    {
+      city: "Nabiganj",
+      sub: "Goplarbazar",
+      postalCode: "3371",
+    },
+    {
+      city: "Nabiganj",
+      sub: "Golduba",
+      postalCode: "3372",
+    },
+    {
+      city: "Nabiganj",
+      sub: "Digalbak",
+      postalCode: "3373",
+    },
+    {
+      city: "Nabiganj",
+      sub: "Inathganj",
+      postalCode: "3374",
+    },
+  ],
+  Kustia: [
+    {
+      city: "Bheramara",
+      sub: "Bheramara",
+      postalCode: "7040",
+    },
+    {
+      city: "Bheramara",
+      sub: "Ganges Bheramara",
+      postalCode: "7041",
+    },
+    {
+      city: "Bheramara",
+      sub: "Allardarga",
+      postalCode: "7042",
+    },
+    {
+      city: "Janipur",
+      sub: "Janipur",
+      postalCode: "7020",
+    },
+    {
+      city: "Janipur",
+      sub: "Khoksa",
+      postalCode: "7021",
+    },
+    {
+      city: "Kumarkhali",
+      sub: "Kumarkhali",
+      postalCode: "7010",
+    },
+    {
+      city: "Kumarkhali",
+      sub: "Panti",
+      postalCode: "7011",
+    },
+    {
+      city: "Kustia Sadar",
+      sub: "Kustia Sadar",
+      postalCode: "7000",
+    },
+    {
+      city: "Kustia Sadar",
+      sub: "Kushtia Mohini",
+      postalCode: "7001",
+    },
+    {
+      city: "Kustia Sadar",
+      sub: "Jagati",
+      postalCode: "7002",
+    },
+    {
+      city: "Kustia Sadar",
+      sub: "Islami University",
+      postalCode: "7003",
+    },
+    {
+      city: "Mirpur",
+      sub: "Mirpur",
+      postalCode: "7030",
+    },
+    {
+      city: "Mirpur",
+      sub: "Poradaha",
+      postalCode: "7031",
+    },
+    {
+      city: "Mirpur",
+      sub: "Amla Sadarpur",
+      postalCode: "7032",
+    },
+    {
+      city: "Rafayetpur",
+      sub: "Rafayetpur",
+      postalCode: "7050",
+    },
+    {
+      city: "Rafayetpur",
+      sub: "Taragunia",
+      postalCode: "7051",
+    },
+    {
+      city: "Rafayetpur",
+      sub: "Khasmathurapur",
+      postalCode: "7052",
+    },
+  ],
+  Meherpur: [
+    {
+      city: "Gangni",
+      sub: "Gangni",
+      postalCode: "7110",
+    },
+    {
+      city: "Meherpur Sadar",
+      sub: "Meherpur Sadar",
+      postalCode: "7100",
+    },
+    {
+      city: "Meherpur Sadar",
+      sub: "Amjhupi",
+      postalCode: "7101",
+    },
+    {
+      city: "Meherpur Sadar",
+      sub: "Mujib Nagar Complex",
+      postalCode: "7102",
+    },
+    {
+      city: "Meherpur Sadar",
+      sub: "Amjhupi",
+      postalCode: "7152",
+    },
+  ],
+  Chuadanga: [
+    {
+      city: "Alamdanga",
+      sub: "Alamdanga",
+      postalCode: "7210",
+    },
+    {
+      city: "Alamdanga",
+      sub: "Hardi",
+      postalCode: "7211",
+    },
+    {
+      city: "Chuadanga Sadar",
+      sub: "Chuadanga Sadar",
+      postalCode: "7200",
+    },
+    {
+      city: "Chuadanga Sadar",
+      sub: "Munshiganj",
+      postalCode: "7201",
+    },
+    {
+      city: "Damurhuda",
+      sub: "Damurhuda",
+      postalCode: "7220",
+    },
+    {
+      city: "Damurhuda",
+      sub: "Darshana",
+      postalCode: "7221",
+    },
+    {
+      city: "Damurhuda",
+      sub: "Andulbaria",
+      postalCode: "7222",
+    },
+    {
+      city: "Doulatganj",
+      sub: "Doulatganj",
+      postalCode: "7230",
+    },
+  ],
+  Jinaidaha: [
+    {
+      city: "Harinakundu",
+      sub: "Harinakundu",
+      postalCode: "7310",
+    },
+    {
+      city: "Jinaidaha Sadar",
+      sub: "Jinaidaha Sadar",
+      postalCode: "7300",
+    },
+    {
+      city: "Jinaidaha Sadar",
+      sub: "Jinaidaha Cadet College",
+      postalCode: "7301",
+    },
+    {
+      city: "Kotchandpur",
+      sub: "Kotchandpur",
+      postalCode: "7330",
+    },
+    {
+      city: "Maheshpur",
+      sub: "Maheshpur",
+      postalCode: "7340",
+    },
+    {
+      city: "Naldanga",
+      sub: "Naldanga",
+      postalCode: "7350",
+    },
+    {
+      city: "Naldanga",
+      sub: "Hatbar Bazar",
+      postalCode: "7351",
+    },
+    {
+      city: "Shailakupa",
+      sub: "Shailakupa",
+      postalCode: "7320",
+    },
+    {
+      city: "Shailakupa",
+      sub: "Kumiradaha",
+      postalCode: "7321",
+    },
+  ],
+  Jessore: [
+    {
+      city: "Bagharpara",
+      sub: "Bagharpara",
+      postalCode: "7470",
+    },
+    {
+      city: "Bagharpara",
+      sub: "Gouranagar",
+      postalCode: "7471",
+    },
+    {
+      city: "Chaugachha",
+      sub: "Chougachha",
+      postalCode: "7410",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Jessore Sadar",
+      postalCode: "7400",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Jessore Upa-Shahar",
+      postalCode: "7401",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Chanchra",
+      postalCode: "7402",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Jessore canttonment",
+      postalCode: "7403",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Jessore Airbach",
+      postalCode: "7404",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Rupdia",
+      postalCode: "7405",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Basundia",
+      postalCode: "7406",
+    },
+    {
+      city: "Jessore Sadar",
+      sub: "Churamankathi",
+      postalCode: "7407",
+    },
+    {
+      city: "Jhikargachha",
+      sub: "Jhikargachha",
+      postalCode: "7420",
+    },
+    {
+      city: "Keshabpur",
+      sub: "Keshobpur",
+      postalCode: "7450",
+    },
+    {
+      city: "Monirampur",
+      sub: "Monirampur",
+      postalCode: "7440",
+    },
+    {
+      city: "Noapara",
+      sub: "Noapara",
+      postalCode: "7460",
+    },
+    {
+      city: "Noapara",
+      sub: "Rajghat",
+      postalCode: "7461",
+    },
+    {
+      city: "Noapara",
+      sub: "Bhugilhat",
+      postalCode: "7462",
+    },
+    {
+      city: "Sarsa",
+      sub: "Sarsa",
+      postalCode: "7430",
+    },
+    {
+      city: "Sarsa",
+      sub: "Benapole",
+      postalCode: "7431",
+    },
+    {
+      city: "Sarsa",
+      sub: "Jadabpur",
+      postalCode: "7432",
+    },
+    {
+      city: "Sarsa",
+      sub: "Bag Achra",
+      postalCode: "7433",
+    },
+  ],
+  Narail: [
+    {
+      city: "Kalia",
+      sub: "Kalia",
+      postalCode: "7520",
+    },
+    {
+      city: "Laxmipasha",
+      sub: "Laxmipasha",
+      postalCode: "7510",
+    },
+    {
+      city: "Laxmipasha",
+      sub: "Lohagora",
+      postalCode: "7511",
+    },
+    {
+      city: "Laxmipasha",
+      sub: "Itna",
+      postalCode: "7512",
+    },
+    {
+      city: "Laxmipasha",
+      sub: "Naldi",
+      postalCode: "7513",
+    },
+    {
+      city: "Laxmipasha",
+      sub: "Baradia",
+      postalCode: "7514",
+    },
+    {
+      city: "Mohajan",
+      sub: "Mohajan",
+      postalCode: "7521",
+    },
+    {
+      city: "Narail Sadar",
+      sub: "Narail Sadar",
+      postalCode: "7500",
+    },
+    {
+      city: "Narail Sadar",
+      sub: "Ratanganj",
+      postalCode: "7501",
+    },
+  ],
+  Magura: [
+    {
+      city: "Arpara",
+      sub: "Arpara",
+      postalCode: "7620",
+    },
+    {
+      city: "Magura Sadar",
+      sub: "Magura Sadar",
+      postalCode: "7600",
+    },
+    {
+      city: "Mohammadpur",
+      sub: "Mohammadpur",
+      postalCode: "7630",
+    },
+    {
+      city: "Mohammadpur",
+      sub: "Binodpur",
+      postalCode: "7631",
+    },
+    {
+      city: "Mohammadpur",
+      sub: "Nahata",
+      postalCode: "7632",
+    },
+    {
+      city: "Shripur",
+      sub: "Shripur",
+      postalCode: "7610",
+    },
+    {
+      city: "Shripur",
+      sub: "Langalbadh",
+      postalCode: "7611",
+    },
+    {
+      city: "Shripur",
+      sub: "Nachol",
+      postalCode: "7612",
+    },
+  ],
+  Khulna: [
+    {
+      city: "Alaipur",
+      sub: "Alaipur",
+      postalCode: "9240",
+    },
+    {
+      city: "Alaipur",
+      sub: "Rupsha",
+      postalCode: "9241",
+    },
+    {
+      city: "Alaipur",
+      sub: "Belphulia",
+      postalCode: "9242",
+    },
+    {
+      city: "Batiaghat",
+      sub: "Batiaghat",
+      postalCode: "9260",
+    },
+    {
+      city: "Batiaghat",
+      sub: "Surkalee",
+      postalCode: "9261",
+    },
+    {
+      city: "Chalna Bazar",
+      sub: "Chalna Bazar",
+      postalCode: "9270",
+    },
+    {
+      city: "Chalna Bazar",
+      sub: "Dakup",
+      postalCode: "9271",
+    },
+    {
+      city: "Chalna Bazar",
+      sub: "Bajua",
+      postalCode: "9272",
+    },
+    {
+      city: "Chalna Bazar",
+      sub: "Nalian",
+      postalCode: "9273",
+    },
+    {
+      city: "Digalia",
+      sub: "Digalia",
+      postalCode: "9220",
+    },
+    {
+      city: "Digalia",
+      sub: "Chandni Mahal",
+      postalCode: "9221",
+    },
+    {
+      city: "Digalia",
+      sub: "Senhati",
+      postalCode: "9222",
+    },
+    {
+      city: "Digalia",
+      sub: "Ghoshghati",
+      postalCode: "9223",
+    },
+    {
+      city: "Digalia",
+      sub: "Gazirhat",
+      postalCode: "9224",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Khulna G.P.O",
+      postalCode: "9000",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Khula Sadar",
+      postalCode: "9100",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Khulna Shipyard",
+      postalCode: "9201",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Doulatpur",
+      postalCode: "9202",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "BIT Khulna",
+      postalCode: "9203",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Siramani",
+      postalCode: "9204",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Jahanabad Canttonmen",
+      postalCode: "9205",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Sonali Jute Mills",
+      postalCode: "9206",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Atra Shilpa Area",
+      postalCode: "9207",
+    },
+    {
+      city: "Khulna Sadar",
+      sub: "Khulna University",
+      postalCode: "9208",
+    },
+    {
+      city: "Madinabad",
+      sub: "Madinabad",
+      postalCode: "9290",
+    },
+    {
+      city: "Madinabad",
+      sub: "Amadee",
+      postalCode: "9291",
+    },
+    {
+      city: "Paikgachha",
+      sub: "Paikgachha",
+      postalCode: "9280",
+    },
+    {
+      city: "Paikgachha",
+      sub: "Godaipur",
+      postalCode: "9281",
+    },
+    {
+      city: "Paikgachha",
+      sub: "Kapilmoni",
+      postalCode: "9282",
+    },
+    {
+      city: "Paikgachha",
+      sub: "Katipara",
+      postalCode: "9283",
+    },
+    {
+      city: "Paikgachha",
+      sub: "Chandkhali",
+      postalCode: "9284",
+    },
+    {
+      city: "Paikgachha",
+      sub: "Garaikhali",
+      postalCode: "9285",
+    },
+    {
+      city: "Phultala",
+      sub: "Phultala",
+      postalCode: "9210",
+    },
+    {
+      city: "Sajiara",
+      sub: "Sajiara",
+      postalCode: "9250",
+    },
+    {
+      city: "Sajiara",
+      sub: "Ghonabanda",
+      postalCode: "9251",
+    },
+    {
+      city: "Sajiara",
+      sub: "Chuknagar",
+      postalCode: "9252",
+    },
+    {
+      city: "Sajiara",
+      sub: "Shahapur",
+      postalCode: "9253",
+    },
+    {
+      city: "Terakhada",
+      sub: "Terakhada",
+      postalCode: "9230",
+    },
+    {
+      city: "Terakhada",
+      sub: "Pak Barasat",
+      postalCode: "9231",
+    },
+  ],
+  "IBH WAs Here": [
+    {
+      city: "Bagerhat Sadar",
+      sub: "Bagerhat Sadar",
+      postalCode: "9300",
+    },
+    {
+      city: "Bagerhat Sadar",
+      sub: "P.C College",
+      postalCode: "9301",
+    },
+    {
+      city: "Bagerhat Sadar",
+      sub: "Rangdia",
+      postalCode: "9302",
+    },
+    {
+      city: "Chalna Ankorage",
+      sub: "Chalna Ankorage",
+      postalCode: "9350",
+    },
+    {
+      city: "Chalna Ankorage",
+      sub: "Mongla Port",
+      postalCode: "9351",
+    },
+    {
+      city: "Chitalmari",
+      sub: "Chitalmari",
+      postalCode: "9360",
+    },
+    {
+      city: "Chitalmari",
+      sub: "Barabaria",
+      postalCode: "9361",
+    },
+    {
+      city: "Fakirhat",
+      sub: "Fakirhat",
+      postalCode: "9370",
+    },
+    {
+      city: "Fakirhat",
+      sub: "Mansa",
+      postalCode: "9371",
+    },
+    {
+      city: "Fakirhat",
+      sub: "Bhanganpar Bazar",
+      postalCode: "9372",
+    },
+    {
+      city: "Kachua UPO",
+      sub: "Kachua",
+      postalCode: "9310",
+    },
+    {
+      city: "Kachua UPO",
+      sub: "Sonarkola",
+      postalCode: "9311",
+    },
+    {
+      city: "Mollahat",
+      sub: "Mollahat",
+      postalCode: "9380",
+    },
+    {
+      city: "Mollahat",
+      sub: "Kahalpur",
+      postalCode: "9381",
+    },
+    {
+      city: "Mollahat",
+      sub: "Dariala",
+      postalCode: "9382",
+    },
+    {
+      city: "Mollahat",
+      sub: "Charkulia",
+      postalCode: "9383",
+    },
+    {
+      city: "Mollahat",
+      sub: "Nagarkandi",
+      postalCode: "9384",
+    },
+    {
+      city: "Mollahat",
+      sub: "Pak Gangni",
+      postalCode: "9385",
+    },
+    {
+      city: "Morelganj",
+      sub: "Morelganj",
+      postalCode: "9320",
+    },
+    {
+      city: "Morelganj",
+      sub: "Sannasi Bazar",
+      postalCode: "9321",
+    },
+    {
+      city: "Morelganj",
+      sub: "Telisatee",
+      postalCode: "9322",
+    },
+    {
+      city: "Rampal",
+      sub: "Rampal",
+      postalCode: "9340",
+    },
+    {
+      city: "Rampal",
+      sub: "Foylahat",
+      postalCode: "9341",
+    },
+    {
+      city: "Rampal",
+      sub: "Sonatunia",
+      postalCode: "9342",
+    },
+    {
+      city: "Rampal",
+      sub: "Gourambha",
+      postalCode: "9343",
+    },
+    {
+      city: "Rayenda",
+      sub: "Rayenda",
+      postalCode: "9330",
+    },
+  ],
+  Satkhira: [
+    {
+      city: "Ashashuni",
+      sub: "Ashashuni",
+      postalCode: "9460",
+    },
+    {
+      city: "Ashashuni",
+      sub: "Baradal",
+      postalCode: "9461",
+    },
+    {
+      city: "Debbhata",
+      sub: "Debbhata",
+      postalCode: "9430",
+    },
+    {
+      city: "Debbhata",
+      sub: "Gurugram",
+      postalCode: "9431",
+    },
+    {
+      city: "kalaroa",
+      sub: "kalaroa",
+      postalCode: "9410",
+    },
+    {
+      city: "kalaroa",
+      sub: "Murarikati",
+      postalCode: "9411",
+    },
+    {
+      city: "kalaroa",
+      sub: "Jhaudanga",
+      postalCode: "9412",
+    },
+    {
+      city: "kalaroa",
+      sub: "Hamidpur",
+      postalCode: "9413",
+    },
+    {
+      city: "kalaroa",
+      sub: "Khordo",
+      postalCode: "9414",
+    },
+    {
+      city: "kalaroa",
+      sub: "Chandanpur",
+      postalCode: "9415",
+    },
+    {
+      city: "Kaliganj UPO",
+      sub: "Kaliganj UPO",
+      postalCode: "9440",
+    },
+    {
+      city: "Kaliganj UPO",
+      sub: "Nalta Mubaroknagar",
+      postalCode: "9441",
+    },
+    {
+      city: "Kaliganj UPO",
+      sub: "Ratanpur",
+      postalCode: "9442",
+    },
+    {
+      city: "Nakipur",
+      sub: "Nakipur",
+      postalCode: "9450",
+    },
+    {
+      city: "Nakipur",
+      sub: "Noornagar",
+      postalCode: "9451",
+    },
+    {
+      city: "Nakipur",
+      sub: "Naobeki",
+      postalCode: "9452",
+    },
+    {
+      city: "Nakipur",
+      sub: "Buri Goalini",
+      postalCode: "9453",
+    },
+    {
+      city: "Nakipur",
+      sub: "Gabura",
+      postalCode: "9454",
+    },
+    {
+      city: "Nakipur",
+      sub: "Habinagar",
+      postalCode: "9455",
+    },
+    {
+      city: "Satkhira Sadar",
+      sub: "Satkhira Sadar",
+      postalCode: "9400",
+    },
+    {
+      city: "Satkhira Sadar",
+      sub: "Satkhira Islamia Acc",
+      postalCode: "9401",
+    },
+    {
+      city: "Satkhira Sadar",
+      sub: "Gunakar kati",
+      postalCode: "9402",
+    },
+    {
+      city: "Satkhira Sadar",
+      sub: "Budhhat",
+      postalCode: "9403",
+    },
+    {
+      city: "Taala",
+      sub: "Taala",
+      postalCode: "9420",
+    },
+    {
+      city: "Taala",
+      sub: "Patkelghata",
+      postalCode: "9421",
+    },
+  ],
+  Brahmanbaria: [
+    {
+      city: "Akhaura",
+      sub: "Akhaura",
+      postalCode: "3450",
+    },
+    {
+      city: "Akhaura",
+      sub: "Azampur",
+      postalCode: "3451",
+    },
+    {
+      city: "Akhaura",
+      sub: "Gangasagar",
+      postalCode: "3452",
+    },
+    {
+      city: "Banchharampur",
+      sub: "Banchharampur",
+      postalCode: "3420",
+    },
+    {
+      city: "Brahamanbaria Sadar",
+      sub: "Brahamanbaria Sadar",
+      postalCode: "3400",
+    },
+    {
+      city: "Brahamanbaria Sadar",
+      sub: "Talshahar",
+      postalCode: "3401",
+    },
+    {
+      city: "Brahamanbaria Sadar",
+      sub: "Ashuganj",
+      postalCode: "3402",
+    },
+    {
+      city: "Brahamanbaria Sadar",
+      sub: "Ashuganj Share",
+      postalCode: "3403",
+    },
+    {
+      city: "Brahamanbaria Sadar",
+      sub: "Poun",
+      postalCode: "3404",
+    },
+    {
+      city: "Kasba",
+      sub: "Kasba",
+      postalCode: "3460",
+    },
+    {
+      city: "Kasba",
+      sub: "Kuti",
+      postalCode: "3461",
+    },
+    {
+      city: "Kasba",
+      sub: "Chandidar",
+      postalCode: "3462",
+    },
+    {
+      city: "Kasba",
+      sub: "Chargachh",
+      postalCode: "3463",
+    },
+    {
+      city: "Kasba",
+      sub: "Gopinathpur",
+      postalCode: "3464",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Nabinagar",
+      postalCode: "3410",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Laubfatehpur",
+      postalCode: "3411",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Rasullabad",
+      postalCode: "3412",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Shamgram",
+      postalCode: "3413",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Ratanpur",
+      postalCode: "3414",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Shahapur",
+      postalCode: "3415",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Kaitala",
+      postalCode: "3417",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Salimganj",
+      postalCode: "3418",
+    },
+    {
+      city: "Nabinagar",
+      sub: "Jibanganj",
+      postalCode: "3419",
+    },
+    {
+      city: "Nasirnagar",
+      sub: "Nasirnagar",
+      postalCode: "3440",
+    },
+    {
+      city: "Nasirnagar",
+      sub: "Fandauk",
+      postalCode: "3441",
+    },
+    {
+      city: "Sarail",
+      sub: "Sarail",
+      postalCode: "3430",
+    },
+    {
+      city: "Sarail",
+      sub: "Shahbajpur",
+      postalCode: "3431",
+    },
+    {
+      city: "Sarail",
+      sub: "Chandura",
+      postalCode: "3432",
+    },
+  ],
+  Comilla: [
+    {
+      city: "Barura",
+      sub: "Barura",
+      postalCode: "3560",
+    },
+    {
+      city: "Barura",
+      sub: "Poyalgachha",
+      postalCode: "3561",
+    },
+    {
+      city: "Barura",
+      sub: "Murdafarganj",
+      postalCode: "3562",
+    },
+    {
+      city: "Brahmanpara",
+      sub: "Brahmanpara",
+      postalCode: "3526",
+    },
+    {
+      city: "Burichang",
+      sub: "Burichang",
+      postalCode: "3520",
+    },
+    {
+      city: "Burichang",
+      sub: "Maynamoti bazar",
+      postalCode: "3521",
+    },
+    {
+      city: "Chandina",
+      sub: "Chandia",
+      postalCode: "3510",
+    },
+    {
+      city: "Chandina",
+      sub: "Madhaiabazar",
+      postalCode: "3511",
+    },
+    {
+      city: "Chouddagram",
+      sub: "Chouddagram",
+      postalCode: "3550",
+    },
+    {
+      city: "Chouddagram",
+      sub: "Batisa",
+      postalCode: "3551",
+    },
+    {
+      city: "Chouddagram",
+      sub: "Chiora",
+      postalCode: "3552",
+    },
+    {
+      city: "Comilla Sadar",
+      sub: "Comilla Sadar",
+      postalCode: "3500",
+    },
+    {
+      city: "Comilla Sadar",
+      sub: "Comilla Contoment",
+      postalCode: "3501",
+    },
+    {
+      city: "Comilla Sadar",
+      sub: "Halimanagar",
+      postalCode: "3502",
+    },
+    {
+      city: "Comilla Sadar",
+      sub: "Courtbari",
+      postalCode: "3503",
+    },
+    {
+      city: "Comilla Sadar",
+      sub: "Suaganj",
+      postalCode: "3504",
+    },
+    {
+      city: "Daudkandi",
+      sub: "Daudkandi",
+      postalCode: "3516",
+    },
+    {
+      city: "Daudkandi",
+      sub: "Gouripur",
+      postalCode: "3517",
+    },
+    {
+      city: "Daudkandi",
+      sub: "Dashpara",
+      postalCode: "3518",
+    },
+    {
+      city: "Daudkandi",
+      sub: "Eliotganj",
+      postalCode: "3519",
+    },
+    {
+      city: "Davidhar",
+      sub: "Davidhar",
+      postalCode: "3530",
+    },
+    {
+      city: "Davidhar",
+      sub: "Gangamandal",
+      postalCode: "3531",
+    },
+    {
+      city: "Davidhar",
+      sub: "Barashalghar",
+      postalCode: "3532",
+    },
+    {
+      city: "Davidhar",
+      sub: "Dhamtee",
+      postalCode: "3533",
+    },
+    {
+      city: "Homna",
+      sub: "Homna",
+      postalCode: "3546",
+    },
+    {
+      city: "Laksam",
+      sub: "Laksam",
+      postalCode: "3570",
+    },
+    {
+      city: "Laksam",
+      sub: "Lakshamanpur",
+      postalCode: "3571",
+    },
+    {
+      city: "Laksam",
+      sub: "Bipulasar",
+      postalCode: "3572",
+    },
+    {
+      city: "Langalkot",
+      sub: "Langalkot",
+      postalCode: "3580",
+    },
+    {
+      city: "Langalkot",
+      sub: "Dhalua",
+      postalCode: "3581",
+    },
+    {
+      city: "Langalkot",
+      sub: "Chhariabazar",
+      postalCode: "3582",
+    },
+    {
+      city: "Langalkot",
+      sub: "Gunabati",
+      postalCode: "3583",
+    },
+    {
+      city: "Muradnagar",
+      sub: "Muradnagar",
+      postalCode: "3540",
+    },
+    {
+      city: "Muradnagar",
+      sub: "Ramchandarpur",
+      postalCode: "3541",
+    },
+    {
+      city: "Muradnagar",
+      sub: "Companyganj",
+      postalCode: "3542",
+    },
+    {
+      city: "Muradnagar",
+      sub: "Bangra",
+      postalCode: "3543",
+    },
+    {
+      city: "Muradnagar",
+      sub: "Sonakanda",
+      postalCode: "3544",
+    },
+    {
+      city: "Muradnagar",
+      sub: "Pantibazar",
+      postalCode: "3545",
+    },
+  ],
+  Lakshmipur: [
+    {
+      city: "Char Alexgander",
+      sub: "Char Alexgander",
+      postalCode: "3730",
+    },
+    {
+      city: "Char Alexgander",
+      sub: "Hajirghat",
+      postalCode: "3731",
+    },
+    {
+      city: "Char Alexgander",
+      sub: "Ramgatirhat",
+      postalCode: "3732",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Lakshimpur Sadar",
+      postalCode: "3700",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Dalal Bazar",
+      postalCode: "3701",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Bhabaniganj",
+      postalCode: "3702",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Mandari",
+      postalCode: "3703",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Keramatganj",
+      postalCode: "3704",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Rupchara",
+      postalCode: "3705",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Duttapara",
+      postalCode: "3706",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Choupalli",
+      postalCode: "3707",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Chandraganj",
+      postalCode: "3708",
+    },
+    {
+      city: "Lakshimpur Sadar",
+      sub: "Amani Lakshimpur",
+      postalCode: "3709",
+    },
+    {
+      city: "Ramganj",
+      sub: "Ramganj",
+      postalCode: "3720",
+    },
+    {
+      city: "Ramganj",
+      sub: "Alipur",
+      postalCode: "3721",
+    },
+    {
+      city: "Ramganj",
+      sub: "Panpara",
+      postalCode: "3722",
+    },
+    {
+      city: "Ramganj",
+      sub: "Kanchanpur",
+      postalCode: "3723",
+    },
+    {
+      city: "Ramganj",
+      sub: "Naagmud",
+      postalCode: "3724",
+    },
+    {
+      city: "Ramganj",
+      sub: "Dolta",
+      postalCode: "3725",
+    },
+    {
+      city: "Raypur",
+      sub: "Raypur",
+      postalCode: "3710",
+    },
+    {
+      city: "Raypur",
+      sub: "Rakhallia",
+      postalCode: "3711",
+    },
+    {
+      city: "Raypur",
+      sub: "Nagerdighirpar",
+      postalCode: "3712",
+    },
+    {
+      city: "Raypur",
+      sub: "Haydarganj",
+      postalCode: "3713",
+    },
+    {
+      city: "Raypur",
+      sub: "Bhuabari",
+      postalCode: "3714",
+    },
+  ],
+  Noakhali: [
+    {
+      city: "Basurhat",
+      sub: "Basur Hat",
+      postalCode: "3850",
+    },
+    {
+      city: "Basurhat",
+      sub: "Charhajari",
+      postalCode: "3851",
+    },
+    {
+      city: "Begumganj",
+      sub: "Begumganj",
+      postalCode: "3820",
+    },
+    {
+      city: "Begumganj",
+      sub: "Choumohani",
+      postalCode: "3821",
+    },
+    {
+      city: "Begumganj",
+      sub: "Banglabazar",
+      postalCode: "3822",
+    },
+    {
+      city: "Begumganj",
+      sub: "Mir Owarishpur",
+      postalCode: "3823",
+    },
+    {
+      city: "Begumganj",
+      sub: "Bazra",
+      postalCode: "3824",
+    },
+    {
+      city: "Begumganj",
+      sub: "Jamidar Hat",
+      postalCode: "3825",
+    },
+    {
+      city: "Begumganj",
+      sub: "Sonaimuri",
+      postalCode: "3827",
+    },
+    {
+      city: "Begumganj",
+      sub: "Gopalpur",
+      postalCode: "3828",
+    },
+    {
+      city: "Begumganj",
+      sub: "Joynarayanpur",
+      postalCode: "3829",
+    },
+    {
+      city: "Begumganj",
+      sub: "Alaiarpur",
+      postalCode: "3831",
+    },
+    {
+      city: "Begumganj",
+      sub: "Tangirpar",
+      postalCode: "3832",
+    },
+    {
+      city: "Begumganj",
+      sub: "Khalafat Bazar",
+      postalCode: "3833",
+    },
+    {
+      city: "Begumganj",
+      sub: "Rajganj",
+      postalCode: "3834",
+    },
+    {
+      city: "Begumganj",
+      sub: "Oachhekpur",
+      postalCode: "3835",
+    },
+    {
+      city: "Begumganj",
+      sub: "Bhabani Jibanpur",
+      postalCode: "3837",
+    },
+    {
+      city: "Begumganj",
+      sub: "Maheshganj",
+      postalCode: "3838",
+    },
+    {
+      city: "Begumganj",
+      sub: "Nadona",
+      postalCode: "3839",
+    },
+    {
+      city: "Begumganj",
+      sub: "Nandiapara",
+      postalCode: "3841",
+    },
+    {
+      city: "Begumganj",
+      sub: "Khalishpur",
+      postalCode: "3842",
+    },
+    {
+      city: "Begumganj",
+      sub: "Dauti",
+      postalCode: "3843",
+    },
+    {
+      city: "Begumganj",
+      sub: "Joyag",
+      postalCode: "3844",
+    },
+    {
+      city: "Begumganj",
+      sub: "Thanar Hat",
+      postalCode: "3845",
+    },
+    {
+      city: "Begumganj",
+      sub: "Amisha Para",
+      postalCode: "3847",
+    },
+    {
+      city: "Begumganj",
+      sub: "Durgapur",
+      postalCode: "3848",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Chatkhil",
+      postalCode: "3870",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Palla",
+      postalCode: "3871",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Khilpara",
+      postalCode: "3872",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Bodalcourt",
+      postalCode: "3873",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Rezzakpur",
+      postalCode: "3874",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Solla",
+      postalCode: "3875",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Karihati",
+      postalCode: "3877",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Dosh Gharia",
+      postalCode: "3878",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Bansa Bazar",
+      postalCode: "3879",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Sahapur",
+      postalCode: "3881",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Sampara",
+      postalCode: "3882",
+    },
+    {
+      city: "Chatkhil",
+      sub: "Shingbahura",
+      postalCode: "3883",
+    },
+    {
+      city: "Hatiya",
+      sub: "Hatiya",
+      postalCode: "3890",
+    },
+    {
+      city: "Hatiya",
+      sub: "Afazia",
+      postalCode: "3891",
+    },
+    {
+      city: "Hatiya",
+      sub: "Tamoraddi",
+      postalCode: "3892",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Noakhali Sadar",
+      postalCode: "3800",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Noakhali College",
+      postalCode: "3801",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Sonapur",
+      postalCode: "3802",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Din Monir Hat",
+      postalCode: "3803",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Pak Kishoreganj",
+      postalCode: "3804",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Mriddarhat",
+      postalCode: "3806",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Kabirhat",
+      postalCode: "3807",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Khalifar Hat",
+      postalCode: "3808",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Charam Tua",
+      postalCode: "3809",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Chaprashir Hat",
+      postalCode: "3811",
+    },
+    {
+      city: "Noakhali Sadar",
+      sub: "Char Jabbar",
+      postalCode: "3812",
+    },
+    {
+      city: "Senbag",
+      sub: "Senbag",
+      postalCode: "3860",
+    },
+    {
+      city: "Senbag",
+      sub: "Kallyandi",
+      postalCode: "3861",
+    },
+    {
+      city: "Senbag",
+      sub: "Beezbag",
+      postalCode: "3862",
+    },
+    {
+      city: "Senbag",
+      sub: "Kankirhat",
+      postalCode: "3863",
+    },
+    {
+      city: "Senbag",
+      sub: "Chatarpaia",
+      postalCode: "3864",
+    },
+    {
+      city: "Senbag",
+      sub: "T.P. Lamua",
+      postalCode: "3865",
+    },
+  ],
+  Feni: [
+    {
+      city: "Chhagalnaia",
+      sub: "Chhagalnaia",
+      postalCode: "3910",
+    },
+    {
+      city: "Chhagalnaia",
+      sub: "Maharajganj",
+      postalCode: "3911",
+    },
+    {
+      city: "Chhagalnaia",
+      sub: "Daraga Hat",
+      postalCode: "3912",
+    },
+    {
+      city: "Chhagalnaia",
+      sub: "Puabashimulia",
+      postalCode: "3913",
+    },
+    {
+      city: "Dagonbhuia",
+      sub: "Dagondhuia",
+      postalCode: "3920",
+    },
+    {
+      city: "Dagonbhuia",
+      sub: "Dudmukha",
+      postalCode: "3921",
+    },
+    {
+      city: "Dagonbhuia",
+      sub: "Chhilonia",
+      postalCode: "3922",
+    },
+    {
+      city: "Dagonbhuia",
+      sub: "Rajapur",
+      postalCode: "3923",
+    },
+    {
+      city: "Feni Sadar",
+      sub: "Feni Sadar",
+      postalCode: "3900",
+    },
+    {
+      city: "Feni Sadar",
+      sub: "Fazilpur",
+      postalCode: "3901",
+    },
+    {
+      city: "Feni Sadar",
+      sub: "Sharshadie",
+      postalCode: "3902",
+    },
+    {
+      city: "Feni Sadar",
+      sub: "Laskarhat",
+      postalCode: "3903",
+    },
+    {
+      city: "Pashurampur",
+      sub: "Pashurampur",
+      postalCode: "3940",
+    },
+    {
+      city: "Pashurampur",
+      sub: "Shuarbazar",
+      postalCode: "3941",
+    },
+    {
+      city: "Pashurampur",
+      sub: "Fulgazi",
+      postalCode: "3942",
+    },
+    {
+      city: "Pashurampur",
+      sub: "Munshirhat",
+      postalCode: "3943",
+    },
+    {
+      city: "Sonagazi",
+      sub: "Sonagazi",
+      postalCode: "3930",
+    },
+    {
+      city: "Sonagazi",
+      sub: "Motiganj",
+      postalCode: "3931",
+    },
+    {
+      city: "Sonagazi",
+      sub: "Ahmadpur",
+      postalCode: "3932",
+    },
+    {
+      city: "Sonagazi",
+      sub: "Kazirhat",
+      postalCode: "3933",
+    },
+  ],
+  Chittagong: [
+    {
+      city: "Anawara",
+      sub: "Anowara",
+      postalCode: "4376",
+    },
+    {
+      city: "Anawara",
+      sub: "Paroikora",
+      postalCode: "4377",
+    },
+    {
+      city: "Anawara",
+      sub: "Battali",
+      postalCode: "4378",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Kanungopara",
+      postalCode: "4363",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Saroatoli",
+      postalCode: "4364",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Iqbal Park",
+      postalCode: "4365",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Boalkhali",
+      postalCode: "4366",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Sakpura",
+      postalCode: "4367",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Kadurkhal",
+      postalCode: "4368",
+    },
+    {
+      city: "Boalkhali",
+      sub: "Charandwip",
+      postalCode: "4369",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chittagong GPO",
+      postalCode: "4000",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chittagong Bandar",
+      postalCode: "4100",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Pahartoli",
+      postalCode: "4202",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chawkbazar",
+      postalCode: "4203",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Patenga",
+      postalCode: "4204",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chittagong Airport",
+      postalCode: "4205",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Jaldia Merine Accade",
+      postalCode: "4206",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Firozshah",
+      postalCode: "4207",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Mohard",
+      postalCode: "4208",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chitt. Politechnic In",
+      postalCode: "4209",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Bayezid Bostami",
+      postalCode: "4210",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Amin Jute Mills",
+      postalCode: "4211",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chandgaon",
+      postalCode: "4212",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Wazedia",
+      postalCode: "4213",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Jalalabad",
+      postalCode: "4214",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Anandabazar",
+      postalCode: "4215",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Halishahar",
+      postalCode: "4216",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "North Katuli",
+      postalCode: "4217",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chitt. Sailers Colon",
+      postalCode: "4218",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chitt. Customs Acca",
+      postalCode: "4219",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Chitt. Cantonment",
+      postalCode: "4220",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Al- Amin Baria Madra",
+      postalCode: "4221",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Middle Patenga",
+      postalCode: "4222",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Export Processing",
+      postalCode: "4223",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Rampura TSO",
+      postalCode: "4224",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "Halishshar",
+      postalCode: "4225",
+    },
+    {
+      city: "Chittagong Sadar",
+      sub: "North Halishahar",
+      postalCode: "4226",
+    },
+    {
+      city: "East Joara",
+      sub: "East Joara",
+      postalCode: "4380",
+    },
+    {
+      city: "East Joara",
+      sub: "Gachbaria",
+      postalCode: "4381",
+    },
+    {
+      city: "East Joara",
+      sub: "Dohazari",
+      postalCode: "4382",
+    },
+    {
+      city: "East Joara",
+      sub: "Barma",
+      postalCode: "4383",
+    },
+    {
+      city: "Fatikchhari",
+      sub: "Fatikchhari",
+      postalCode: "4350",
+    },
+    {
+      city: "Fatikchhari",
+      sub: "Nanupur",
+      postalCode: "4351",
+    },
+    {
+      city: "Fatikchhari",
+      sub: "Bhandar Sharif",
+      postalCode: "4352",
+    },
+    {
+      city: "Fatikchhari",
+      sub: "Najirhat",
+      postalCode: "4353",
+    },
+    {
+      city: "Fatikchhari",
+      sub: "Harualchhari",
+      postalCode: "4354",
+    },
+    {
+      city: "Fatikchhari",
+      sub: "Narayanhat",
+      postalCode: "4355",
+    },
+    {
+      city: "Hathazari",
+      sub: "Hathazari",
+      postalCode: "4330",
+    },
+    {
+      city: "Hathazari",
+      sub: "Chitt.University",
+      postalCode: "4331",
+    },
+    {
+      city: "Hathazari",
+      sub: "Gorduara",
+      postalCode: "4332",
+    },
+    {
+      city: "Hathazari",
+      sub: "Katirhat",
+      postalCode: "4333",
+    },
+    {
+      city: "Hathazari",
+      sub: "Mirzapur",
+      postalCode: "4334",
+    },
+    {
+      city: "Hathazari",
+      sub: "Fatahabad",
+      postalCode: "4335",
+    },
+    {
+      city: "Hathazari",
+      sub: "Nuralibari",
+      postalCode: "4337",
+    },
+    {
+      city: "Hathazari",
+      sub: "Yunus Nagar",
+      postalCode: "4338",
+    },
+    {
+      city: "Hathazari",
+      sub: "Madrasa",
+      postalCode: "4339",
+    },
+    {
+      city: "Jaldi",
+      sub: "Jaldi",
+      postalCode: "4390",
+    },
+    {
+      city: "Jaldi",
+      sub: "Khan Bahadur",
+      postalCode: "4391",
+    },
+    {
+      city: "Jaldi",
+      sub: "Gunagari",
+      postalCode: "4392",
+    },
+    {
+      city: "Jaldi",
+      sub: "Banigram",
+      postalCode: "4393",
+    },
+    {
+      city: "Lohagara",
+      sub: "Lohagara",
+      postalCode: "4396",
+    },
+    {
+      city: "Lohagara",
+      sub: "Padua",
+      postalCode: "4397",
+    },
+    {
+      city: "Lohagara",
+      sub: "Chunti",
+      postalCode: "4398",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Mirsharai",
+      postalCode: "4320",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Abutorab",
+      postalCode: "4321",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Darrogahat",
+      postalCode: "4322",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Bharawazhat",
+      postalCode: "4323",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Joarganj",
+      postalCode: "4324",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Azampur",
+      postalCode: "4325",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Korerhat",
+      postalCode: "4327",
+    },
+    {
+      city: "Mirsharai",
+      sub: "Mohazanhat",
+      postalCode: "4328",
+    },
+    {
+      city: "Patia Head Office",
+      sub: "Patia Head Office",
+      postalCode: "4370",
+    },
+    {
+      city: "Patia Head Office",
+      sub: "Budhpara",
+      postalCode: "4371",
+    },
+    {
+      city: "Rangunia",
+      sub: "Rangunia",
+      postalCode: "4360",
+    },
+    {
+      city: "Rangunia",
+      sub: "Dhamair",
+      postalCode: "4361",
+    },
+    {
+      city: "Rouzan",
+      sub: "Rouzan",
+      postalCode: "4340",
+    },
+    {
+      city: "Rouzan",
+      sub: "Beenajuri",
+      postalCode: "4341",
+    },
+    {
+      city: "Rouzan",
+      sub: "Kundeshwari",
+      postalCode: "4342",
+    },
+    {
+      city: "Rouzan",
+      sub: "Gahira",
+      postalCode: "4343",
+    },
+    {
+      city: "Rouzan",
+      sub: "jagannath Hat",
+      postalCode: "4344",
+    },
+    {
+      city: "Rouzan",
+      sub: "Fatepur",
+      postalCode: "4345",
+    },
+    {
+      city: "Rouzan",
+      sub: "Guzra Noapara",
+      postalCode: "4346",
+    },
+    {
+      city: "Rouzan",
+      sub: "Dewanpur",
+      postalCode: "4347",
+    },
+    {
+      city: "Rouzan",
+      sub: "Mohamuni",
+      postalCode: "4348",
+    },
+    {
+      city: "Rouzan",
+      sub: "B.I.T Post Office",
+      postalCode: "4349",
+    },
+    {
+      city: "Sandwip",
+      sub: "Sandwip",
+      postalCode: "4300",
+    },
+    {
+      city: "Sandwip",
+      sub: "Shiberhat",
+      postalCode: "4301",
+    },
+    {
+      city: "Sandwip",
+      sub: "Urirchar",
+      postalCode: "4302",
+    },
+    {
+      city: "Satkania",
+      sub: "Satkania",
+      postalCode: "4386",
+    },
+    {
+      city: "Satkania",
+      sub: "Baitul Ijjat",
+      postalCode: "4387",
+    },
+    {
+      city: "Satkania",
+      sub: "Bazalia",
+      postalCode: "4388",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Sitakunda",
+      postalCode: "4310",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Baroidhala",
+      postalCode: "4311",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Barabkunda",
+      postalCode: "4312",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Bawashbaria",
+      postalCode: "4313",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Kumira",
+      postalCode: "4314",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Bhatiari",
+      postalCode: "4315",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Fouzdarhat",
+      postalCode: "4316",
+    },
+    {
+      city: "Sitakunda",
+      sub: "Jafrabad",
+      postalCode: "4317",
+    },
+  ],
+  Khagrachari: [
+    {
+      city: "Diginala",
+      sub: "Diginala",
+      postalCode: "4420",
+    },
+    {
+      city: "Khagrachari Sadar",
+      sub: "Khagrachari Sadar",
+      postalCode: "4400",
+    },
+    {
+      city: "Laxmichhari",
+      sub: "Laxmichhari",
+      postalCode: "4470",
+    },
+    {
+      city: "Mahalchhari",
+      sub: "Mahalchhari",
+      postalCode: "4430",
+    },
+    {
+      city: "Manikchhari",
+      sub: "Manikchhari",
+      postalCode: "4460",
+    },
+    {
+      city: "Matiranga",
+      sub: "Matiranga",
+      postalCode: "4450",
+    },
+    {
+      city: "Panchhari",
+      sub: "Panchhari",
+      postalCode: "4410",
+    },
+    {
+      city: "Ramghar Head Office",
+      sub: "Ramghar Head Office",
+      postalCode: "4440",
+    },
+  ],
+  Rangamati: [
+    {
+      city: "Barakal",
+      sub: "Barakal",
+      postalCode: "4570",
+    },
+    {
+      city: "Bilaichhari",
+      sub: "Bilaichhari",
+      postalCode: "4550",
+    },
+    {
+      city: "Jarachhari",
+      sub: "Jarachhari",
+      postalCode: "4560",
+    },
+    {
+      city: "Kalampati",
+      sub: "Kalampati",
+      postalCode: "4510",
+    },
+    {
+      city: "Kalampati",
+      sub: "Betbunia",
+      postalCode: "4511",
+    },
+    {
+      city: "kaptai",
+      sub: "Kaptai",
+      postalCode: "4530",
+    },
+    {
+      city: "kaptai",
+      sub: "Chandraghona",
+      postalCode: "4531",
+    },
+    {
+      city: "kaptai",
+      sub: "Kaptai Project",
+      postalCode: "4532",
+    },
+    {
+      city: "kaptai",
+      sub: "Kaptai Nuton Bazar",
+      postalCode: "4533",
+    },
+    {
+      city: "Longachh",
+      sub: "Longachh",
+      postalCode: "4580",
+    },
+    {
+      city: "Marishya",
+      sub: "Marishya",
+      postalCode: "4590",
+    },
+    {
+      city: "Naniachhar",
+      sub: "Nanichhar",
+      postalCode: "4520",
+    },
+    {
+      city: "Rajsthali",
+      sub: "Rajsthali",
+      postalCode: "4540",
+    },
+    {
+      city: "Rangamati Sadar",
+      sub: "Rangamati Sadar",
+      postalCode: "4500",
+    },
+  ],
+  Bandarban: [
+    {
+      city: "Alikadam",
+      sub: "Alikadam",
+      postalCode: "4650",
+    },
+    {
+      city: "Bandarban Sadar",
+      sub: "Bandarban Sadar",
+      postalCode: "4600",
+    },
+    {
+      city: "Naikhong",
+      sub: "Naikhong",
+      postalCode: "4660",
+    },
+    {
+      city: "Roanchhari",
+      sub: "Roanchhari",
+      postalCode: "4610",
+    },
+    {
+      city: "Ruma",
+      sub: "Ruma",
+      postalCode: "4620",
+    },
+    {
+      city: "Thanchi",
+      sub: "Thanchi",
+      postalCode: "4630",
+    },
+    {
+      city: "Thanchi",
+      sub: "Lama",
+      postalCode: "4641",
+    },
+  ],
+  "Coxs Bazar": [
+    {
+      city: "Chiringga",
+      sub: "Chiringga",
+      postalCode: "4740",
+    },
+    {
+      city: "Chiringga",
+      sub: "Chiringga S.O",
+      postalCode: "4741",
+    },
+    {
+      city: "Chiringga",
+      sub: "Badarkali",
+      postalCode: "4742",
+    },
+    {
+      city: "Chiringga",
+      sub: "Malumghat",
+      postalCode: "4743",
+    },
+    {
+      city: "Coxs Bazar Sadar",
+      sub: "Coxs Bazar Sadar",
+      postalCode: "4700",
+    },
+    {
+      city: "Coxs Bazar Sadar",
+      sub: "Zhilanja",
+      postalCode: "4701",
+    },
+    {
+      city: "Coxs Bazar Sadar",
+      sub: "Eidga",
+      postalCode: "4702",
+    },
+    {
+      city: "Gorakghat",
+      sub: "Gorakghat",
+      postalCode: "4710",
+    },
+    {
+      city: "Kutubdia",
+      sub: "Kutubdia",
+      postalCode: "4720",
+    },
+    {
+      city: "Ramu",
+      sub: "Ramu",
+      postalCode: "4730",
+    },
+    {
+      city: "Teknaf",
+      sub: "Teknaf",
+      postalCode: "4760",
+    },
+    {
+      city: "Teknaf",
+      sub: "Hnila",
+      postalCode: "4761",
+    },
+    {
+      city: "Teknaf",
+      sub: "St.Martin",
+      postalCode: "4762",
+    },
+    {
+      city: "Ukhia",
+      sub: "Ukhia",
+      postalCode: "4750",
+    },
+  ],
+};
+
+module.exports = data;
